==> 2_FIGURES/00_data_wrangling.Rmd <==
---
title: "OO_data_wrangling"
output: html_document
date: "2024-04-03"
editor_options: 
  chunk_output_type: console
  markdown: 
    wrap: 72
---

#LIBRARY 

```{r}
library(tidyverse)
library(googlesheets4)
library(oce)
library(padr)
library(rEDM)
library(ggplot2)
library(zoo)
library(quantreg)

```

#List of all data

## Daily (ag_df_CR)

```{r}

#CR BUOY

ag_df_CR

ag_df_CR$chl

ag_df_CR$phy

ag_df_CR$ph

ag_df_CR$Celsius

ag_df_CR$spcond

ag_df_CR$do

ag_df_CR$turb

ag_df_CR$Salinity

#ag_Logan
                          
    ag_Logan$HourlyAltimeterSetting 

    ag_Logan$HourlyDewPointTemperature 
    
    ag_Logan$HourlyDryBulbTemperature 
    
    ag_Logan$HourlyPrecipitation 
    
   ag_Logan$HourlyRelativeHumidity 
   
    ag_Logan$HourlyStationPressure 
    
    ag_Logan$HourlyVisibility
    
    ag_Logan$HourlyWetBulbTemperature 
    
    ag_Logan$HourlyWindDirection 
    
    ag_Logan$HourlyWindGustSpeed
    
    ag_Logan$HourlyWindSpeed 
    
    #CAM017
    
    CAM017$MG
    
    #bacteria 
    
    ag_CR_bacteria$Enterococcus
    ag_CR_bacteria$E_coli
    ag_CR_nutrients$`Ammonium (uM)`
    
    #nutrients 
    ag_CR_nutrients$`Nitrate+nitrite (uM)`
    ag_CR_nutrients$`Phosphate (uM)`
    ag_CR_nutrients$`Phosphate (uM)`
    ag_CR_nutrients$`Total phosphorus (uM)`
    ag_CR_nutrients$`Chlorophyll a (ug/L)`
    ag_CR_nutrients$`Phaeophytin (ug/L)`
    
#aggregated
  
All.ag_Logan
All.CR_nutrients
All.CR_bacteria 


All.Var 

    
```

## 9 AM - Noon (morndf_CR)

```{r}

morndf_CR

morndf_CR$chl

mornddf_CR$phy

mornddf_CR$ph

mornddf_CR$Celsius

mornddf_CR$spcond

mornddf_CR$do

mornddf_CR$turb

mornddf_CR$Salinity


```

## 15 Min (CR_15)

```{r}

CR_15

CR_15$chl

CR_15$phy

CR_15$ph

CR_15$Celsius

CR_15$spcond

CR_15$do

CR_15$turb

CR_15$Salinity


```

##LOGAN (ag_Logan)
 
```{r}

    ag_Logan$HourlyAltimeterSetting 

    ag_Logan$HourlyDewPointTemperature 
    
    ag_Logan$HourlyDryBulbTemperature 
    
    ag_Logan$HourlyPrecipitation 
    
   ag_Logan$HourlyRelativeHumidity 
   
    ag_Logan$HourlyStationPressure 
    
    ag_Logan$HourlyVisibility
    
    ag_Logan$HourlyWetBulbTemperature 
    
    ag_Logan$HourlyWindDirection 
    
    ag_Logan$HourlyWindGustSpeed
    
    ag_Logan$HourlyWindSpeed 
    
  


```

##MWRA Non-CSO

```{r}

#ag_CR_secchi

ag_CR_secchi$`Total Suspended Solids (mg/L)`

#ag_CR_nutrients 

ag_CR_nutrients$`Ammonium (uM)`
ag_CR_nutrients$`Nitrate+nitrite (uM)`
ag_CR_nutrients$`Phosphate (uM)`
ag_CR_nutrients$`Total phosphorus (uM)`
ag_CR_nutrients$`Phaeophytin (ug/L)`

#ag_CR_bacteria

ag_CR_bacteria$Enterococcus
ag_CR_bacteria$E_coli
ag_CR_bacteria$coliform


```

##CSO

```{r}

CAM017$MG

```

```{r}

All.CR_bacteria

All.CR_secchi

All.CR_nutrients 

```

# 15 MIN (CR_15)

## Load EPA Buoy Data

```{r}

library(googlesheets4)

# Define Google Sheets ID map
google_sheets_map <- list(
  `2015` = "1t5joquxku9OfQGDdsabp-GdR2sUw8hPDcwatC3SL90Q",
  `2016` = "1sMcEOIzxPTUZ93-pTJBE0lQ3SI8kKihLZRdpZ32_Lyw",
  `2017` = "11SO-RxRhcc4qqn3gRt4n13JWClFoG--kISMoPAR2FRA",
  `2018` = "1CO0DHuKb3C-hE9z-r5yR-SXwPnZWZg4vQDuEtRv2r7k",
  `2019` = "1PGZCChA-8HU-cLvPSxqrxd89gLTletDYgogMGwv4Dg4",
  `2020` = "1Qvt0PIK4wPFPKbkM98pSQp9TbPsu_a6frTWFKv3aNQ4",
  `2021` = "1vaDoymaAQmNFKqIgo1LStI7GQSKDeIqLy7We85IA3Oo",
  `2022` = "1F5l-1dWi7EFOaEKf9o3eUBi6BgQOrC1Rj_lVNRPrx5U",
  `2023` = "1sMKGzW71qErP8DIrxWDvlgqN3o9MrprLjjAFGsOie3I"
)

# Loop through each year and load data
for (year in 2015:2023) {

  message("Loading data for year: ", year)

  # Define file paths
  local_path <- sprintf("C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/CR%d.rds", year)
  github_url <- sprintf("https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/CR%d.rds", year)
  sheet_id <- google_sheets_map[[as.character(year)]]
  google_sheets_url <- sprintf("https://docs.google.com/spreadsheets/d/%s/edit?usp=sharing", sheet_id)

  # Attempt to load data
  assign(paste0("CR", year),
         tryCatch({
           # Try GitHub with proper connection management
           con <- url(github_url)
           on.exit(close(con), add = TRUE)
           data <- readRDS(con)
           saveRDS(data, local_path)
           data
         }, error = function(e1) {
           # Try local file
           if (file.exists(local_path)) {
             readRDS(local_path)
           } else {
             # Fallback to Google Sheets
             data <- read_sheet(google_sheets_url, sheet = 2)
             saveRDS(data, local_path)
             data
           }
         })
  )
}

# Clean up temporary variables
rm(local_path, github_url, sheet_id, google_sheets_url, google_sheets_map, year, data, con)


#closeAllConnections()


```

## Wrangle EPA Buoy Data

```{r}

library(dplyr)
library(lubridate)


#bind rows

CR_15 <- bind_rows(
  CR2015,
  CR2016,
  CR2017, 
  CR2018,
  CR2019,
  CR2020,
  CR2021,
  CR2022,
  CR2023
)


CR_15 <- CR_15 %>%
  mutate(hour = pmin(hour(`time est`),hour(`time edt`),na.rm=T),
         minute = pmin(minute(`time est`),minute(`time edt`),na.rm=T)) %>%
  select(date,hour,minute,chl=`chlorophyll (rfu)`,phy=`phycocyanin (rfu)`, ph=`ph`, Celsius =`temp c`,  spcond = `spcond (ms/cm)`, do = `do (mg/l)`,turb = `turbidity (fnu)` )



missing_CR_15 <- CR_15 %>%
  filter(
    # Keep rows where at least one non-hour, non-minute column has an NA value
    if_any(-c(hour, minute, date), is.na) &
    # Remove rows where all non-hour, non-minute columns are NA
    !if_all(-c(hour, minute, date), is.na)
  )


CR_15 <- na.omit(CR_15)


```

# Add seasonality 365 day cycle (sin_season)

```{r}

library(padr)

CR_15 <- bind_rows(
  CR2015,
  CR2016,
  CR2017, 
  CR2018,
  CR2019,
  CR2020,
  CR2021,
  CR2022,
  CR2023
)


CR_15 <- CR_15 %>%
  mutate(hour = pmin(hour(`time est`),hour(`time edt`),na.rm=T),
         minute = pmin(minute(`time est`),minute(`time edt`),na.rm=T)) %>%
  select(date,hour,minute,chl=`chlorophyll (rfu)`,phy=`phycocyanin (rfu)`, ph=`ph`, Celsius =`temp c`,  spcond = `spcond (ms/cm)`, do = `do (mg/l)`,turb = `turbidity (fnu)` )


CR_15 <- na.omit(CR_15)

CR_15 <- pad(CR_15)

CR_15$DateTime <- seq(
  from = as.POSIXct("2015-05-13 00:00:00"),
  by = "15 min",
  length.out = nrow(CR_15)
)



# Add day of year + fraction of the day
CR_15 <- CR_15 %>%
  mutate(
    yday_frac = yday(DateTime) + 
                (hour(DateTime) + minute(DateTime)/60) / 24,
    sin_season = sin(2 * pi * yday_frac / 365)
  )

sum(is.na(CR_15$sin_season))  # Count NA values in sin_season

CR_15 <- na.omit(CR_15)


CR_15 <- CR_15 %>%
  select(-DateTime, -yday_frac)

```


# Add practical salinity (Salinity)

##Convert Salinity m/s to S/m

```{r}

CR_15$spcond = CR_15$spcond * 0.1 #mS to S/m

```

## Calculate practical salinity

```{r}

library(oce)

CR_15$Salinity <- 

  swSCTp(
  CR_15$spcond,
CR_15$Celsius,
  pressure = rep(10.1325,NROW(CR_15)),
  # "S/cm",
"",
  eos = getOption("oceEOS", default = "gsw")
)

```

### Load if not save

```{r}


# CONFIG: Define object name and paths
quoted_file_name <- "CR_15"
rds_name <- paste0(quoted_file_name, ".rds")

local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  rds_name
)

github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  rds_name
)


# Load from GitHub or local, if not save
if (!file.exists(local_path)) {
  tryCatch({
    con <- url(github_url)
    obj <- readRDS(con)
    close(con)  # Close the connection
    assign(quoted_file_name, obj)
    saveRDS(obj, local_path)
    message("Loaded from GitHub and saved locally.")
  }, error = function(e) {
    if (exists(quoted_file_name)) {
      saveRDS(get(quoted_file_name), local_path)
      message("GitHub failed. Saved in-memory object to local path.")
    } else {
      stop(paste0("Failed to load ", quoted_file_name, " from GitHub and no in-memory object exists to save."))
    }
  })
} else {
  assign(quoted_file_name, readRDS(local_path))
  message("Loaded from local path.")
}

# Clean up
rm(rds_name, quoted_file_name, local_path, github_url)


```


## HOURLY (CR_HOURLY)
```{r}

 CR_hourly <- CR_15 %>%
  group_by(date, hour) %>%
  summarise(
    chl = mean(chl, na.rm = TRUE),
    phy = mean(phy, na.rm = TRUE),
    ph = mean(ph, na.rm = TRUE),
    Celsius = mean(Celsius, na.rm = TRUE),
    spcond = mean(spcond, na.rm = TRUE),
    do = mean(do, na.rm = TRUE),
    turb = mean(turb, na.rm = TRUE),
    sin_season = mean(sin_season, na.rm = TRUE),
    Salinity = mean(Salinity, na.rm = TRUE),
    .groups = "drop"
  )



# CONFIG: Define object name and paths
quoted_file_name <- "CR_hourly"
rds_name <- paste0(quoted_file_name, ".rds")

local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  rds_name
)

github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  rds_name
)


# Load from GitHub or local, if not save
if (!file.exists(local_path)) {
  tryCatch({
    con <- url(github_url)
    obj <- readRDS(con)
    close(con)  # Close the connection
    assign(quoted_file_name, obj)
    saveRDS(obj, local_path)
    message("Loaded from GitHub and saved locally.")
  }, error = function(e) {
    if (exists(quoted_file_name)) {
      saveRDS(get(quoted_file_name), local_path)
      message("GitHub failed. Saved in-memory object to local path.")
    } else {
      stop(paste0("Failed to load ", quoted_file_name, " from GitHub and no in-memory object exists to save."))
    }
  })
} else {
  assign(quoted_file_name, readRDS(local_path))
  message("Loaded from local path.")
}

# Clean up
rm(rds_name, quoted_file_name, local_path, github_url)




```


#plot misaligned values  

```{r}
library(dplyr)
library(tidyr)
library(ggplot2)


CR_15 <- CR_15 %>%
  mutate(hour = pmin(hour(`time est`),hour(`time edt`),na.rm=T),
         minute = pmin(minute(`time est`),minute(`time edt`),na.rm=T)) %>%
  select(date,hour,minute,chl=`chlorophyll (rfu)`,phy=`phycocyanin (rfu)`, ph=`ph`, Celsius =`temp c`,  spcond = `spcond (ms/cm)`, do = `do (mg/l)`,turb = `turbidity (fnu)` )



missing_CR_15 <- CR_15 %>%
  filter(
    # Keep rows where at least one non-hour, non-minute column has an NA value
    if_any(-c(hour, minute, date), is.na) &
    # Remove rows where all non-hour, non-minute columns are NA
    !if_all(-c(hour, minute, date), is.na)
  )


# Replace NA values with 5.555 in the specified columns
missing_CR_15 <- missing_CR_15 %>%
  mutate(across(c(chl, phy, ph, Celsius, spcond, do, turb), ~ replace(., is.na(.), 5.555)))

# Add Source identifiers
missing_CR_15$Source <- "Missing"
CR_15$Source <- "Present"

# Combine both datasets
combined_df <- bind_rows(CR_15, missing_CR_15)

# Ensure Date column is in correct format
combined_df$Date <- as.Date(combined_df$date)

# Reshape to long format
combined_long_df <- combined_df %>%
  pivot_longer(cols = c("chl", "phy", "ph", "Celsius", "spcond", "do", "turb"),
               names_to = "Variable",
               values_to = "Value") %>%
  mutate(Variable = recode(Variable,
    "chl" = "Chlorophyll (RFU)",
    "phy" = "Phycocyanin (RFU)",
    "ph" = "pH",
    "Celsius" = "Temperature (°C)",
    "spcond" = "Specific Conductivity (mS/cm)",
    "turb" = "Turbidity (FNU)",
    "do" = "Dissolved Oxygen (mg/L)")
  ) %>%
  # Assign color categories based on value and source
  mutate(Value_Color = case_when(
    Value == 5.555 ~ "highlight",                   # 5.555 values
    Source == "Missing" ~ "missing_but_real_value", # Missing but not 5.555
    TRUE ~ "normal"                                 # All other values
  ))

# Plot
ggplot(combined_long_df, aes(x = Date, y = Variable, fill = Value_Color)) +
  geom_tile(height = 0.8) +
  scale_fill_manual(values = c(
    "normal" = "darkblue",
    "NA Value" = "orange",
    "Value Misaligned with NA" = "lightgreen"
  )) +
  labs(
    x = "Date",
    y = "Variable",
    title = "Data Timeline (Present vs Missing vs 5.555 Values)"
  ) +
  theme_minimal() +
  theme(
    axis.text.y = element_text(size = 8),
    plot.title = element_text(hjust = 0.5)
  )


```


```{r}
library(dplyr)
library(tidyr)
library(ggplot2)

# Replace NA values with 5.555 in the specified columns
missing_CR_15 <- missing_CR_15 %>%
  mutate(across(c(chl, phy, ph, Celsius, spcond, do, turb), ~ replace(., is.na(.), 5.555)))

# Add Source identifiers
missing_CR_15$Source <- "Missing"
CR_15$Source <- "Present"

# Combine both datasets
combined_df <- bind_rows(CR_15, missing_CR_15)

# Ensure Date column is in correct format
combined_df$Date <- as.Date(combined_df$date)

# Reshape to long format
combined_long_df <- combined_df %>%
  pivot_longer(cols = c("chl", "phy", "ph", "Celsius", "spcond", "do", "turb"),
               names_to = "Variable",
               values_to = "Value") %>%
  mutate(Variable = recode(Variable,
    "chl" = "Chlorophyll (RFU)",
    "phy" = "Phycocyanin (RFU)",
    "ph" = "pH",
    "Celsius" = "Temperature (°C)",
    "spcond" = "Specific Conductivity (mS/cm)",
    "turb" = "Turbidity (FNU)",
    "do" = "Dissolved Oxygen (mg/L)")
  ) %>%
  # Assign color categories based on value and source
  mutate(Value_Color = case_when(
    Value == 5.555 ~ "Value Misaligned with NA",   # Highlighted replacement value
    Source == "Missing" ~ "NA Value",              # Originally missing but not 5.555
    TRUE ~ "normal"                                # All other values
  ))

# Plot
ggplot(combined_long_df, aes(x = Date, y = Variable, fill = Value_Color)) +
  geom_tile(height = 0.8) +
  scale_fill_manual(values = c(
    "normal" = "darkblue",
    "NA Value" = "orange",
    "Value Misaligned with NA" = "lightgreen"
  )) +
  labs(
    x = "Date",
    y = "Variable",
    title = "Data Timeline (Present vs Missing vs 5.555 Values)",
    fill = "Data Type"
  ) +
  theme_minimal() +
  theme(
    axis.text.y = element_text(size = 8),
    plot.title = element_text(hjust = 0.5)
  )

```



### "Load, if not Save" 15 min Data (CR_15)

```{r}


# CONFIG: Define object name and paths
quoted_file_name <- "CR_15"
rds_name <- paste0(quoted_file_name, ".rds")

local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  rds_name
)

github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  rds_name
)


# Load from GitHub or local, if not save
if (!file.exists(local_path)) {
  tryCatch({
    con <- url(github_url)
    obj <- readRDS(con)
    close(con)  # Close the connection
    assign(quoted_file_name, obj)
    saveRDS(obj, local_path)
    message("Loaded from GitHub and saved locally.")
  }, error = function(e) {
    if (exists(quoted_file_name)) {
      saveRDS(get(quoted_file_name), local_path)
      message("GitHub failed. Saved in-memory object to local path.")
    } else {
      stop(paste0("Failed to load ", quoted_file_name, " from GitHub and no in-memory object exists to save."))
    }
  })
} else {
  assign(quoted_file_name, readRDS(local_path))
  message("Loaded from local path.")
}

# Clean up
rm(rds_name, quoted_file_name, local_path, github_url)




```


#Daily (ag_df_CR)


```{r}



# CONFIG: Define object name and paths
quoted_file_name <- "CR_15"
rds_name <- paste0(quoted_file_name, ".rds")

local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  rds_name
)

github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  rds_name
)


# Load from GitHub or local, if not save
if (!file.exists(local_path)) {
  tryCatch({
    con <- url(github_url)
    obj <- readRDS(con)
    close(con)  # Close the connection
    assign(quoted_file_name, obj)
    saveRDS(obj, local_path)
    message("Loaded from GitHub and saved locally.")
  }, error = function(e) {
    if (exists(quoted_file_name)) {
      saveRDS(get(quoted_file_name), local_path)
      message("GitHub failed. Saved in-memory object to local path.")
    } else {
      stop(paste0("Failed to load ", quoted_file_name, " from GitHub and no in-memory object exists to save."))
    }
  })
} else {
  assign(quoted_file_name, readRDS(local_path))
  message("Loaded from local path.")
}

# Clean up
rm(rds_name, quoted_file_name, local_path, github_url)




#ag_df_CR


ag_df_CR <-CR_15 %>%
  mutate(date = trunc(date,"days")) %>%
   group_by(date) %>%
  summarise(chl=mean(chl, na.rm = TRUE), phy=mean(phy, na.rm = TRUE), ph=mean(ph, na.rm = TRUE), Celsius=mean(Celsius,na.rm = TRUE), spcond=mean(spcond,na.rm = TRUE), do=mean(do,na.rm = TRUE), turb=mean(turb,na.rm = TRUE), Salinity = mean(Salinity,na.rm = TRUE), sin_season = mean(sin_season,na.rm = TRUE))


sum(is.na(ag_df_CR$phy))  # Count NA values

ag_df_CR$date <- as.Date(ag_df_CR$date)





# CONFIG: Define object name and paths
quoted_file_name <- "ag_df_CR"
rds_name <- paste0(quoted_file_name, ".rds")

local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  rds_name
)

github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  rds_name
)


# Load from GitHub or local, if not save
if (!file.exists(local_path)) {
  tryCatch({
    con <- url(github_url)
    obj <- readRDS(con)
    close(con)  # Close the connection
    assign(quoted_file_name, obj)
    saveRDS(obj, local_path)
    message("Loaded from GitHub and saved locally.")
  }, error = function(e) {
    if (exists(quoted_file_name)) {
      saveRDS(get(quoted_file_name), local_path)
      message("GitHub failed. Saved in-memory object to local path.")
    } else {
      stop(paste0("Failed to load ", quoted_file_name, " from GitHub and no in-memory object exists to save."))
    }
  })
} else {
  assign(quoted_file_name, readRDS(local_path))
  message("Loaded from local path.")
}

# Clean up
rm(rds_name, quoted_file_name, local_path, github_url)






```


# 9 AM - 12 PM (morndf_CR)

```{r}
# Load CR_15 from local file if it exists, otherwise try GitHub

# Define file name and paths
file_name <- "CR_15.rds"
local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  file_name
)
github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  file_name
)

# Attempt to load from local, else GitHub
CR_15 <- tryCatch({
  if (file.exists(local_path)) {
    message("Loading CR_15 from local path.")
    readRDS(local_path)
  } else {
    message("Local file not found. Attempting to load CR_15 from GitHub.")
    con <- url(github_url)
    on.exit(close(con), add = TRUE)  # Ensure connection is closed
    data <- readRDS(con)
    saveRDS(data, local_path)
    message("Loaded from GitHub and saved locally.")
    data
  }
}, error = function(e) {
  stop("Failed to load CR_15 from both local and GitHub.")
})


# Clean up
rm(file_name, local_path, github_url)

morndf_CR <- CR_15 %>%
  filter(hour %in% c(9, 10, 11))

morndf_CR <-morndf_CR %>% 
   mutate(date = trunc(date,"days")) %>%
   group_by(date) %>%
   summarise(chl=mean(chl, na.rm = TRUE), phy=mean(phy, na.rm = TRUE), ph=mean(ph, na.rm = TRUE), Celsius=mean(Celsius,na.rm = TRUE), spcond=mean(spcond,na.rm = TRUE), do=mean(do,na.rm = TRUE), turb=mean(turb,na.rm = TRUE), Salinity = mean(Salinity,na.rm = TRUE), sin_season = mean(sin_season,na.rm = TRUE))

sum(is.na(morndf_CR$phy))  # Count NA values

morndf_CR$date <- as.Date(morndf_CR$date)


# CONFIG: Define object name and paths
quoted_file_name <- "morndf_CR"
rds_name <- paste0(quoted_file_name, ".rds")

local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  rds_name
)

github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  rds_name
)


# Load from GitHub or local, if not save
if (!file.exists(local_path)) {
  tryCatch({
    con <- url(github_url)
    obj <- readRDS(con)
    close(con)  # Close the connection
    assign(quoted_file_name, obj)
    saveRDS(obj, local_path)
    message("Loaded from GitHub and saved locally.")
  }, error = function(e) {
    if (exists(quoted_file_name)) {
      saveRDS(get(quoted_file_name), local_path)
      message("GitHub failed. Saved in-memory object to local path.")
    } else {
      stop(paste0("Failed to load ", quoted_file_name, " from GitHub and no in-memory object exists to save."))
    }
  })
} else {
  assign(quoted_file_name, readRDS(local_path))
  message("Loaded from local path.")
}

# Clean up
rm(rds_name, quoted_file_name, local_path, github_url)




```

###LOGAN

```{r}

Logan <- read_sheet('https://docs.google.com/spreadsheets/d/1h3mNXsGZxB4fhkZQGr2e8z6CxAQtumA5a208-GEYvBw/edit?usp=sharing')

# Convert to POSIXct object
Logan$DATE <- as.POSIXct(Logan$DATE, format = "%Y-%m-%dT%H:%M:%S")

Logan2 <- Logan

# Remove "s" suffixes from HourlyAltimeterSetting column
Logan2$HourlyAltimeterSetting <- gsub("s$", "", Logan2$HourlyAltimeterSetting)
Logan2$HourlySeaLevelPressure <- gsub("s$", "", Logan2$HourlySeaLevelPressure)
Logan2$HourlyDewPointTemperature <- gsub("s$", "", Logan2$HourlyDewPointTemperature)
Logan2$HourlyPrecipitation <- gsub("T", "0", Logan2$HourlyPrecipitation)


Logan2 <- Logan2 %>%
  mutate_at(vars(HourlyAltimeterSetting, HourlyDewPointTemperature, HourlyDryBulbTemperature,
                 HourlyPrecipitation, HourlyPressureChange, HourlyPressureTendency,
                 HourlyRelativeHumidity, HourlySeaLevelPressure, HourlyStationPressure,
                 HourlyVisibility, HourlyWetBulbTemperature, HourlyWindDirection,
                 HourlyWindGustSpeed, HourlyWindSpeed), ~ifelse(. == "NULL", NA, .))

# Convert columns to numeric (excluding DATE)
numeric_cols <- c("HourlyAltimeterSetting", "HourlyDewPointTemperature", "HourlyDryBulbTemperature",
                  "HourlyPrecipitation", "HourlyPressureChange", "HourlyPressureTendency",
                  "HourlyRelativeHumidity", "HourlySeaLevelPressure", "HourlyStationPressure",
                  "HourlyVisibility", "HourlyWetBulbTemperature", "HourlyWindDirection",
                  "HourlyWindGustSpeed", "HourlyWindSpeed")

Logan2 <- Logan2 %>%
  mutate(across(all_of(numeric_cols), as.numeric))


ag_Logan <- Logan2 %>%
  group_by(DATE) %>%
  summarise(
    HourlyAltimeterSetting = mean(HourlyAltimeterSetting, na.rm = TRUE),
    HourlyDewPointTemperature = mean(HourlyDewPointTemperature, na.rm = TRUE),
    HourlyDryBulbTemperature = mean(HourlyDryBulbTemperature, na.rm = TRUE), 
    HourlyPrecipitation = sum(HourlyPrecipitation, na.rm = TRUE),
   # HourlyPressureChange = mean(HourlyPressureChange, na.rm = TRUE), SKIP
   # HourlyPressureTendency = mean(HourlyPressureTendency, na.rm = TRUE), SKIP
    HourlyRelativeHumidity = mean(HourlyRelativeHumidity, na.rm = TRUE), 
   # HourlySeaLevelPressure = mean(HourlySeaLevelPressure, na.rm = TRUE), SKIP
    HourlyStationPressure = mean(HourlyStationPressure, na.rm = TRUE),
    HourlyVisibility = mean(HourlyVisibility, na.rm = TRUE),
    HourlyWetBulbTemperature = mean(HourlyWetBulbTemperature, na.rm = TRUE), 
    HourlyWindDirection = mean(HourlyWindDirection, na.rm = TRUE), 
    HourlyWindGustSpeed = mean(HourlyWindGustSpeed, na.rm = TRUE), 
    HourlyWindSpeed = mean(HourlyWindSpeed, na.rm = TRUE)
    
    )
  mutate(date = trunc(date,"days")) %>%


# Convert to POSIXct object
ag_Logan$DATE <- as.POSIXct(ag_Logan$DATE, format = "%Y-%m-%dT%H:%M:%S")

ag_Logan <-  ag_Logan %>%
   mutate(DATE = as.Date(DATE)) %>%
  group_by(DATE) %>%
  summarise(
    HourlyAltimeterSetting = mean(HourlyAltimeterSetting, na.rm = TRUE),
    HourlyDewPointTemperature = mean(HourlyDewPointTemperature, na.rm = TRUE),
    HourlyDryBulbTemperature = mean(HourlyDryBulbTemperature, na.rm = TRUE), 
    HourlyPrecipitation = sum(HourlyPrecipitation, na.rm = TRUE),
   # HourlyPressureChange = mean(HourlyPressureChange, na.rm = TRUE), SKIP
   # HourlyPressureTendency = mean(HourlyPressureTendency, na.rm = TRUE), SKIP
    HourlyRelativeHumidity = mean(HourlyRelativeHumidity, na.rm = TRUE), 
   # HourlySeaLevelPressure = mean(HourlySeaLevelPressure, na.rm = TRUE), SKIP
    HourlyStationPressure = mean(HourlyStationPressure, na.rm = TRUE),
    HourlyVisibility = mean(HourlyVisibility, na.rm = TRUE),
    HourlyWetBulbTemperature = mean(HourlyWetBulbTemperature, na.rm = TRUE), 
    HourlyWindDirection = mean(HourlyWindDirection, na.rm = TRUE), 
    HourlyWindGustSpeed = mean(HourlyWindGustSpeed, na.rm = TRUE), 
    HourlyWindSpeed = mean(HourlyWindSpeed, na.rm = TRUE)
    
    )


if (file.exists("ag_Logan.rds")) {
  # If it's saved locally, load the data
  ag_Logan <- readRDS("ag_Logan.rds")
} else {
  saveRDS(ag_Logan, "ag_Logan.rds")
}

#ag_Logan <- na.omit(ag_Logan)



```

### MWRA Non-CSO

```{r}
#READ DATA IN

#CR_bacteria

if (file.exists("CR_bacteria.rds")) {
  # If it's saved locally, load the data
  CR_bacteria <- readRDS("CR_bacteria.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CR_bacteria <- read_sheet('https://docs.google.com/spreadsheets/d/1V1DJc4er2gctuRN2bv3u3Db0bGxz566ec8j6xmXFZ70/edit?usp=sharing', range = 'cr_bacteria')  
saveRDS(CR_bacteria, "CR_bacteria.rds")
}

#CR_secchi


if (file.exists("CR_secchi.rds")) {
  # If it's saved locally, load the data
 CR_secchi <- readRDS("CR_secchi.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CR_secchi <- read_sheet('https://docs.google.com/spreadsheets/d/1fW4_60i-SYkdTPlFQQdN0b5nlZirc2uQyjDN_dvWA18/edit?usp=sharing', range = 'cr_secchi')
saveRDS(CR_secchi, "CR_secchi.rds")
}


#CR_physical 

if (file.exists("CR_physical.rds")) {
  # If it's saved locally, load the data
 CR_physical <- readRDS("CR_physical.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CR_physical <- read_sheet('https://docs.google.com/spreadsheets/d/1P5kzPHFABPTKffxM2yQBTIQWq9NAmnGd3PF6kOsSFNU/edit?usp=sharing', range = 'cr_physical')
saveRDS(CR_physical, "CR_physical.rds")
}

#CR_nutrients

if (file.exists("CR_nutrients.rds")) {
  # If it's saved locally, load the data
 CR_nutrients <- readRDS("CR_nutrients.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CR_nutrients <- read_sheet('https://docs.google.com/spreadsheets/d/1kPHmgJUHF7GtNJJDme_gWgB7hTS7i6SE7pqazChKMZ4/edit?usp=sharing')
saveRDS(CR_physical, "CR_nutrients.rds")
}



```

```{r}
#NEW NAMES 
## CRBacteria 
Bacteria_new_names <- c(
  "Project_ID",
  "Region",
  "Subregion",
  "DEP_Segment",
  "Station ID",
  "Date",
  "Depth_category",
  "Sample_depth",
  "Enterococcus",
  "Enterococcus_condition",
  "E_coli",
  "E_coli_condition",
  "coliform",
  "coliform_condition",
  "Rainfall_1d",
  "Rainfall_2d",
  "Rainfall_3d"
)

#CRSECCHI
Secchi_new_names <- c(
  "Project ID",
"Region",
"Subregion",
"DEP Segment",
"Station ID",
"Surface or Bottom",
"Depth to bottom (m)",
"Sample depth for TSS (m)",
"Date",
"Total Suspended Solids (mg/L)",
"Total Suspended Solids '<' or '>'",
"Secchi disk depth (m)"
)


#CRPHYSICAL
Physical_new_names <- c(
"Project ID",
"Region",
"Subregion",
"DEP segment",
"Station ID",
"Surface or Bottom",
"Date",
"Depth of measurement (m)",
"Temperature (C)",
"Salinity (PSU)",
"Specific Conductance (mS/cm)",
"Dissolved Oxygen (mg/L)",
"DO Pct Saturation (%)",
"pH",
"Turbidity (NTU)")



#CRNUTRIENTS
Nutrients_new_names <- c(

"Project ID",
"Region",
"Subregion",
"DEP segment",
"Station ID",
"Date",
"Surface or Bottom",
"Sample depth (m)",
"Ammonium (uM)",
"Ammonium '<' or '>'",
"Nitrate+nitrite (uM)",
"Nitrate+nitrite  '<' or '>'",
"Total dissolved N (uM)",
"Total dissolved N  '<' or '>'",
"Particulate nitrogen (uM)",
"Particulate nitrogen  '<' or '>'",
"Total Kjeldahl Nitrogen (uM)",
"TKN '<' or '>'",
"Phosphate (uM)",
"Phosphate '<' or '>'",
"Total dissolved P (uM)",
"Total dissolved P '<' or '>'",
"Particulate P (uM)",
"Particulate P '<' or '>'",
"Total phosphorus (uM)",
"Total phosphorus '<' or '>'",
"Particulate carbon (uM)",
"Particulate carbon     '<' or '>'",
"Chlorophyll a (ug/L)",
"Chlorophyll '<' or '>'",
"Phaeophytin (ug/L)",
"Phaeophytin '<' or '>'"
)

#Initial Data Wrangle 

#CRBACTERIA
#colnames(CR_bacteria) <- as.character(CR_bacteria[5,]) #set header
#CR_bacteria <- CR_bacteria[-c(1:5), ] #Remove 1st 5 Rows
#rename columns

CR_bacteria <- CR_bacteria %>% 
  set_names(Bacteria_new_names) %>%
  mutate(Date = map_vec(Date,as.Date))


#CR SECCHI

#colnames(CR_secchi) <- as.character(CR_secchi[5,]) #set header
#CR_secchi <- CR_secchi[-c(1:5), ] #Remove 1st 5 Rows
#rename columns

CR_secchi <- CR_secchi %>% 
  set_names(Secchi_new_names) %>%
  mutate(Date = map_vec(Date,as.Date))

#CRPHYSICAL

#colnames(CR_physical) <- as.character(CR_physical[5,]) #set header
#CR_physical <- CR_physical[-c(1:5), ] #Remove 1st 5 Rows
#rename columns

CR_physical <- CR_physical %>% 
  set_names(Physical_new_names) %>%
  mutate(Date = map_vec(Date,as.Date))

#CRNUTRIENTS
#colnames(CR_nutrients) <- as.character(CR_nutrients[5,]) #set header
#CR_nutrients <- CR_nutrients[-c(1:5), ] #Remove 1st 5 Rows
#rename columns

CR_nutrients <- CR_nutrients %>% 
  set_names(Nutrients_new_names) %>%
  mutate(Date = map_vec(Date,as.Date))


```

```{r}

#SITE SELECTION

#CR_bacteria
CR_bacteria_011 <- subset(CR_bacteria, CR_bacteria$`Station ID` == "011")


#CR_nutrients

CR_nutrients_166 <- subset(CR_nutrients, CR_nutrients$`Station ID` == "166")

#CR_physical
CR_physical_011 <- subset(CR_physical, CR_physical$`Station ID` == "011")


#CR_secchi

CR_secchi_011 <- subset(CR_secchi, CR_secchi$`Station ID` == "011")

#AGGREGATE DAILY

#CR_bacteria

ag_CR_bacteria <-CR_bacteria_011 %>%
   group_by(Date) %>%
  summarise(Enterococcus = mean(Enterococcus, na.rm = TRUE), E_coli = mean(E_coli, na.rm = TRUE),)
ag_CR_bacteria <- na.omit(ag_CR_bacteria)

if (file.exists("ag_CR_bacteria.rds")) {
  # If it's saved locally, load the data
ag_CR_bacteria <- readRDS("ag_CR_bacteria.rds")
} else {
saveRDS(ag_CR_bacteria, "ag_CR_bacteria.rds")
}

ag_CR_bacteria <- na.omit(ag_CR_bacteria)

#CR_nutrients

ag_CR_nutrients <- CR_nutrients_166 %>%
   group_by(Date) %>%
  summarise(`Ammonium (uM)` = mean(`Ammonium (uM)`, na.rm = TRUE), `Nitrate+nitrite (uM)` = mean(`Nitrate+nitrite (uM)`, na.rm = TRUE), `Phosphate (uM)` = mean(`Phosphate (uM)`, na.rm = TRUE), `Total phosphorus (uM)` = mean(`Total phosphorus (uM)`,na.rm = TRUE), `Chlorophyll a (ug/L)` = mean(`Chlorophyll a (ug/L)`,na.rm = TRUE), `Phaeophytin (ug/L)` = mean(`Phaeophytin (ug/L)`,na.rm = TRUE)) 


if (file.exists("ag_CR_nutrients.rds")) {
  # If it's saved locally, load the data
ag_CR_nutrients <- readRDS("ag_CR_nutrients.rds")
} else {
saveRDS(ag_CR_nutrients, "ag_CR_bacteria.rds")
  
}

ag_CR_nutrients <- na.omit(ag_CR_nutrients)

#CR_physical

ag_CR_physical <- CR_physical_011 %>%
  group_by(Date) %>%
  summarise(`Temperature (C)` = mean(`Temperature (C)`, na.rm = TRUE),
            `Salinity (PSU)` = mean(`Salinity (PSU)`, na.rm = TRUE),
            `Specific Conductance (mS/cm)` = mean(`Specific Conductance (mS/cm)`, na.rm = TRUE),
            `Dissolved Oxygen (mg/L)` = mean(`Dissolved Oxygen (mg/L)`, na.rm = TRUE),
            `DO Pct Saturation (%)` = mean(`DO Pct Saturation (%)`, na.rm = TRUE),
            `pH` = mean(`pH`, na.rm = TRUE),
            `Turbidity (NTU)` = mean(`Turbidity (NTU)`, na.rm = TRUE))



if (file.exists("ag_CR_physical.rds")) {
  # If it's saved locally, load the data
ag_CR_physical<- readRDS("ag_CR_physical.rds")
} else {
saveRDS(ag_CR_physical, "ag_CR_physical.rds")
  
}

ag_CR_physical <- na.omit(ag_CR_physical)


#CR_secchi



ag_CR_secchi <-CR_secchi_011 %>%
   group_by(Date) %>%
  summarise(`Total Suspended Solids (mg/L)`= mean(`Total Suspended Solids (mg/L)`, na.rm = TRUE ), `Secchi disk depth (m)` = mean(`Secchi disk depth (m)`, na.rm = TRUE))


if (file.exists("ag_CR_secchi.rds")) {
  # If it's saved locally, load the data
ag_CR_secchil<- readRDS("ag_CR_secchi.rds")
} else {
saveRDS(ag_CR_secchi, "ag_CR_secchi.rds")
  
}

ag_CR_secchi <- na.omit(ag_CR_secchi)




```

###MWRA CSO

```{r}

#MWRACSOData2016

if (file.exists("MWRACSOData2016.rds")) {
  # If it's saved locally, load the data
MWRACSOData2016 <- readRDS("MWRACSOData2016.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
MWRACSOData2016 <- read_sheet('https://docs.google.com/spreadsheets/d/1BF5lMyYrnILoS-rySwTGILiaqYfxA5-Pz5P2xUmOIq8/edit?usp=sharing')
saveRDS(MWRACSOData2016, "MWRACSOData2016.rds")
}

#MWRACSOData

if (file.exists("MWRACSOData.rds")) {
  # If it's saved locally, load the data
MWRACSOData <- readRDS("MWRACSOData.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
MWRACSOData <- read_sheet('https://docs.google.com/spreadsheets/d/1PRu1yvIuxOnMyazAVkXGv7dkOKUClNxl7CrvrctPREU/edit?usp=sharing')
saveRDS(MWRACSOData, "MWRACSOData.rds")
}

```

```{r}


#Extract Date
MWRACSOData$Date <- as.Date(MWRACSOData$`Start Time`, format = "%m/%d/%Y")


#Remove Excess Charecters

#Remove "Treated" in Volume Column
MWRACSOData$`Volume (MG)` <- gsub("Treated","",as.character(MWRACSOData$`Volume (MG)`))

#Remove "min" in Minutes Column

MWRACSOData$Minutes <- gsub("min","",as.character(MWRACSOData$Minutes))

#Remove Rows w/ "*"
MWRACSOData <- MWRACSOData %>% filter(!grepl("\\*", MWRACSOData$Duration))

#Replace NA w/ 0
MWRACSOData$Minutes[is.na(MWRACSOData$Minutes)] <- 0

#Make Duration Column
MWRACSOData$Minutes <- as.numeric(MWRACSOData$Minutes)
MWRACSOData$Hours <- as.numeric(MWRACSOData$Hours)

MWRACSOData$CombinedMinutes <- MWRACSOData$Minutes + (MWRACSOData$Hours * 60)

#Remove Columns 

MWRACSOData <- MWRACSOData[-c(4:10)]

#Check Dates

unique_dates <- unique(MWRACSOData$Date)


# Sanity Check: CombinedMinutes to numeric and Date to Date format
MWRACSOData <- MWRACSOData %>%
  mutate(CombinedMinutes = as.numeric(CombinedMinutes),
         Date = as.Date(Date))

MWRACSOData <- MWRACSOData %>%
  mutate(`Volume (MG)` = as.numeric(`Volume (MG)`),
         Date = as.Date(Date))

# Group by Date and calculate the sum of CombinedMinutes



MWRACSOData2 <- 
aggregate(MWRACSOData$CombinedMinutes, by=list(MWRACSOData$Date, MWRACSOData$`Outfall Location`), FUN = sum) 


MWRACSOData3 <- 
aggregate(MWRACSOData$`Volume (MG)`, by=list(MWRACSOData$Date, MWRACSOData$`Outfall Location`), FUN = sum) 

MWRACSOData20162 <- 
aggregate(MWRACSOData2016$`Duration (Min)`, by=list(MWRACSOData2016$Date, MWRACSOData2016$CSOName), FUN = sum) 

MWRACSOData20163 <- 
aggregate(MWRACSOData2016$`Volume (MG)`, by=list(MWRACSOData2016$Date, MWRACSOData2016$CSOName), FUN = sum) 


# Rename a column using colnames

colnames(MWRACSOData2)[1:3] <- c("Date","Outfall Location", "Total Minutes")

colnames(MWRACSOData20162)[1:3] <- c("Date","Outfall Location", "Total Minutes")

colnames(MWRACSOData3)[1:3] <- c("Date","Outfall Location", "Volume")

colnames(MWRACSOData20163)[1:3] <- c("Date","Outfall Location", "Volume")

#Set Columns to Be The Same

MWRACSOData <- MWRACSOData[-c(1,4,5,7)]

MWRACSOData2016 <- MWRACSOData2016[-c(3:8)]




#Make New Names

MWRA_new_names <- c(
  "CSO_ID",
  "Outfall Location",
  "Date"
)

MWRA2016_new_names <- c(
  "CSO_ID",
  "Outfall Location",
  "Date"
)

MWRACSOData <- MWRACSOData %>% 
  set_names(MWRA_new_names)
  
MWRACSOData2016 <- MWRACSOData2016 %>% 
  set_names(MWRA2016_new_names)

#Fix Date Format

MWRACSOData$Date <- format(as.Date(MWRACSOData$Date, format=
                           "%Y-%m-%d"          ), "%m-%d-%Y")

#Merge Columns 

MergedMWRACSO <- rbind(MWRACSOData,MWRACSOData2016)

#Order Date

MergedMWRACSO[order(as.Date(MergedMWRACSO$Date, format="%m/%d/%Y")),]

#Fix Date Formats of Future Joins
MWRACSOData2$Date <- format(as.Date(MWRACSOData2$Date, format=
                           "%Y-%m-%d"          ), "%m-%d-%Y")


MWRACSOData3$Date <- format(as.Date(MWRACSOData3$Date, format=
                           "%Y-%m-%d"          ), "%m-%d-%Y")

# Perform the joins

MergedMWRACSO <- merge(MergedMWRACSO, MWRACSOData2[, c("Date", "Outfall Location", "Total Minutes")], 
                       by.x = c("Date", "Outfall Location"), 
                       by.y = c("Date", "Outfall Location"), all.x = TRUE)

MergedMWRACSO <- merge(MergedMWRACSO, MWRACSOData20162[, c("Date", "Outfall Location", "Total Minutes")], 
                       by.x = c("Date", "Outfall Location"), 
                       by.y = c("Date", "Outfall Location"), all.x = TRUE)

# Combine the columns using coalesce
MergedMWRACSO <- MergedMWRACSO %>%
  mutate(TotalMinutes_combined = coalesce(`Total Minutes.x`, `Total Minutes.y`))


# Drop the individual columns if you no longer need them
MergedMWRACSO <- MergedMWRACSO %>%
  select(-`Total Minutes.x`, -`Total Minutes.y`)


MergedMWRACSO <- merge(MergedMWRACSO, MWRACSOData3[, c("Date", "Outfall Location", "Volume")], 
                       by.x = c("Date", "Outfall Location"), 
                       by.y = c("Date", "Outfall Location"), all.x = TRUE)

MergedMWRACSO <- merge(MergedMWRACSO, MWRACSOData20163[, c("Date", "Outfall Location", "Volume")], 
                       by.x = c("Date", "Outfall Location"), 
                       by.y = c("Date", "Outfall Location"), all.x = TRUE)

MergedMWRACSO <- MergedMWRACSO %>%
  mutate(Volume_combined = coalesce(`Volume.x`, `Volume.y`))


# Drop the individual columns if you no longer need them
MergedMWRACSO <- MergedMWRACSO %>%
  select(-`Volume.x`, -`Volume.y`)


```

```{r}

MWRACSOCoords <- read_sheet('https://docs.google.com/spreadsheets/d/1O9RONCW0lGEsPrt-srZU8M4xeNxDM3hJmkA1IWL2mW4/edit?usp=sharing')

MWRACSOCoords2 <- read_sheet('https://docs.google.com/spreadsheets/d/18PK0CcX6Ye5RZJtVUI8mmiaRwLIpRNIzGCwOupSqf-k/edit?usp=sharing')


```

```{r}

MWRACSOFinal <-  
left_join(MergedMWRACSO , MWRACSOCoords ,  by = c("CSO_ID" = "OUTFALL"  ))

#MWRACSOCoords2$OUTFALL_ID <- as.character(MWRACSOCoords2$OUTFALL_ID )

#MWRACSOFinal <-  
#left_join(MergedMWRACSO , MWRACSOCoords2 ,  by = c("CSO_ID" = "OUTFALL_ID"  ))
```

```{r}

if (file.exists("MWRACSOFinal.rds")) {
  # If it's saved locally, load the data
MWRACSOFinal <- readRDS("MWRACSOFinal.rds")
} else {
saveRDS(MWRACSOFinal, "MWRACSOFinal.rds")
}


```

```{r}
# List all objects in the environment
all_objects <- ls()

# Identify objects to keep
objects_to_keep <- c("MWRACSOFinal")

# Identify objects to remove
objects_to_remove <- setdiff(all_objects, objects_to_keep)

# Remove unwanted objects
rm(list = objects_to_remove)

rm(all_objects)
rm(objects_to_keep)
rm(objects_to_remove)

```

###CAMCSO1

```{r}

#CAM2015CSO

if (file.exists("CAM2015CSO.rds")) {
  # If it's saved locally, load the data
CAM2015CSO <- readRDS("CAM2015CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2015CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2015CAMCSO')
saveRDS(CAM2015CSO, "CAM2015CSO.rds")
}


#CAM2016CSO

if (file.exists("CAM2016CSO.rds")) {
  # If it's saved locally, load the data
CAM2016CSO <- readRDS("CAM2016CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2016CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2016CAMCSO')
saveRDS(CAM2016CSO , "CAM2016CSO.rds")
}

CAM2016CSO <- subset(CAM2016CSO, select = c("Date", "CAM017* (MG)"))

#CAM2017CSO

if (file.exists("CAM2017CSO.rds")) {
  # If it's saved locally, load the data
CAM2017CSO <- readRDS("CAM2017CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2017CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2017CAMCSO')
saveRDS(CAM2017CSO , "CAM2017CSO.rds")
}

CAM2017CSO <- subset(CAM2017CSO, select = c("Date", "CAM017* (MG)"))


#CAM2018CSO

if (file.exists("CAM2018CSO.rds")) {
  # If it's saved locally, load the data
CAM2018CSO <- readRDS("CAM2018CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2018CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2018CAMCSO')
saveRDS(CAM2018CSO , "CAM2018CSO.rds")
}



CAM2018CSO <- subset(CAM2018CSO, select = c("Date", "CAM017* (MG)"))

#CAM2019CSO

if (file.exists("CAM2019CSO.rds")) {
  # If it's saved locally, load the data
CAM2019CSO <- readRDS("CAM2019CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2019CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2019CAMCSO')
saveRDS(CAM2019CSO , "CAM2019CSO.rds")
}

CAM2019CSO <- subset(CAM2019CSO, select = c("Date", "CAM017* (MG)"))


#CAM2020CSO

if (file.exists("CAM2020CSO.rds")) {
  # If it's saved locally, load the data
CAM2020CSO <- readRDS("CAM2020CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2020CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2020CAMCSO')
saveRDS(CAM2020CSO , "CAM2020CSO.rds")
}

CAM2020CSO <- subset(CAM2020CSO, select = c("Date", "CAM017* (MG)"))



#CAM2021CSO

if (file.exists("CAM2021CSO.rds")) {
  # If it's saved locally, load the data
CAM2021CSO <- readRDS("CAM2021CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2021CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2021CAMCSO')
saveRDS(CAM2021CSO , "CAM2021CSO.rds")
}

CAM2021CSO <- subset(CAM2021CSO, select = c("Date", "CAM017* (MG)"))

#CAM2022CSO

if (file.exists("CAM2022CSO.rds")) {
  # If it's saved locally, load the data
CAM2022CSO <- readRDS("CAM2022CSO.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
CAM2022CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2022CAMCSO')
saveRDS(CAM2022CSO , "CAM2022CSO.rds")
}

CAM2022CSO <- subset(CAM2022CSO, select = c("Date", "CAM017* (MG)"))


```

```{r}
# Combine all values into a single vector and sort them

library(dplyr)

# Ensure column names are consistent across all data frames
colnames(CAM2015CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2016CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2017CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2018CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2019CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2020CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2021CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2022CSO) <- c("Date", "CAM017* (MG)")

CAM2015CSO$`CAM017* (MG)` <- as.numeric(CAM2015CSO$`CAM017* (MG)`)
CAM2016CSO$`CAM017* (MG)` <- as.numeric(CAM2016CSO$`CAM017* (MG)`)
CAM2017CSO$`CAM017* (MG)` <- as.numeric(CAM2017CSO$`CAM017* (MG)`)
CAM2018CSO$`CAM017* (MG)` <- as.numeric(CAM2018CSO$`CAM017* (MG)`)
CAM2019CSO$`CAM017* (MG)` <- as.numeric(CAM2019CSO$`CAM017* (MG)`)
CAM2020CSO$`CAM017* (MG)` <- as.numeric(CAM2020CSO$`CAM017* (MG)`)
CAM2021CSO$`CAM017* (MG)` <- as.numeric(CAM2021CSO$`CAM017* (MG)`)
CAM2022CSO$`CAM017* (MG)` <- as.numeric(CAM2022CSO$`CAM017* (MG)`)


# Make sure that the "CAM017* (MG)" column is of type double


CAM017 <- bind_rows(
 CAM2016CSO,
  CAM2017CSO, 
  CAM2018CSO,
  CAM2019CSO,
  CAM2020CSO,
  CAM2021CSO,
  CAM2022CSO,
)


CAM017 <- CAM017 %>% rename(MG = `CAM017* (MG)`)

CAM017_discharges <- subset(CAM017, MG != 0)



if (file.exists("CAM017.rds")) {
  # If it's saved locally, load the data
  CAM017 <- readRDS("All.CAM017.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
  saveRDS(All.CAM017, "CAM017.rds")
} 
  


```


###CAMCSO2

```{r}

CAM2015CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2015CAMCSO')


CAM2016CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2016CAMCSO')

CAM2016CSO <- subset(CAM2016CSO, select = c("Date", "CAM017* (MG)"))



CAM2017CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2017CAMCSO')


CAM2017CSO <- subset(CAM2017CSO, select = c("Date", "CAM017* (MG)"))




CAM2018CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2018CAMCSO')


CAM2018CSO <- subset(CAM2018CSO, select = c("Date", "CAM017* (MG)"))

CAM2019CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2019CAMCSO')



CAM2019CSO <- subset(CAM2019CSO, select = c("Date", "CAM017* (MG)"))

CAM2020CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2020CAMCSO')


CAM2020CSO <- subset(CAM2020CSO, select = c("Date", "CAM017* (MG)"))

CAM2021CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2021CAMCSO')


CAM2021CSO <- subset(CAM2021CSO, select = c("Date", "CAM017* (MG)"))

CAM2022CSO <- read_sheet('https://docs.google.com/spreadsheets/d/11v15YsnizQBilaTK-mzJWz-qp1PLJGqQwrxQ2AV7WdA/edit?usp=sharing', sheet = '2022CAMCSO')

CAM2022CSO <- subset(CAM2022CSO, select = c("Date", "CAM017* (MG)"))


```

```{r}
# Combine all values into a single vector and sort them

library(dplyr)

# Ensure column names are consistent across all data frames
colnames(CAM2015CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2016CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2017CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2018CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2019CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2020CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2021CSO) <- c("Date", "CAM017* (MG)")
colnames(CAM2022CSO) <- c("Date", "CAM017* (MG)")

CAM2015CSO$`CAM017* (MG)` <- as.numeric(CAM2015CSO$`CAM017* (MG)`)
CAM2016CSO$`CAM017* (MG)` <- as.numeric(CAM2016CSO$`CAM017* (MG)`)
CAM2017CSO$`CAM017* (MG)` <- as.numeric(CAM2017CSO$`CAM017* (MG)`)
CAM2018CSO$`CAM017* (MG)` <- as.numeric(CAM2018CSO$`CAM017* (MG)`)
CAM2019CSO$`CAM017* (MG)` <- as.numeric(CAM2019CSO$`CAM017* (MG)`)
CAM2020CSO$`CAM017* (MG)` <- as.numeric(CAM2020CSO$`CAM017* (MG)`)
CAM2021CSO$`CAM017* (MG)` <- as.numeric(CAM2021CSO$`CAM017* (MG)`)
CAM2022CSO$`CAM017* (MG)` <- as.numeric(CAM2022CSO$`CAM017* (MG)`)


# Make sure that the "CAM017* (MG)" column is of type double


CAM017 <- bind_rows(
 CAM2016CSO,
  CAM2017CSO, 
  CAM2018CSO,
  CAM2019CSO,
  CAM2020CSO,
  CAM2021CSO,
  CAM2022CSO,
)


CAM017 <- CAM017 %>% rename(MG = `CAM017* (MG)`)

CAM017_discharges <- subset(CAM017, MG != 0)



```

\### ALL CSOs

```{r}

ag_df_CR$Date <- ag_df_CR$date

ag_df_CR <- ag_df_CR[,-c(1) ]

#rm(CR2023, CR2022, CR2020, CR2021, CR2020, CR2019, CR2018, CR2017, CR2016, CR2015, df_CR)

ag_df_CR2 <- ag_df_CR


#Fix Date Formats of Future Joins
MWRACSOFinal$Date <- format(as.Date(MWRACSOFinal$Date, format=
                           "%m-%d-%Y"         ), "%m-%d-%Y")


ag_df_CR2$Date <- format(as.Date(ag_df_CR2$Date , format=
                           "%Y-%m-%d"          ), "%m-%d-%Y")


#join

All <- merge(MWRACSOFinal, ag_df_CR2, by = "Date", all = TRUE)

# Assuming your date column is named "Date"
All$Date <- as.Date(All$Date, format="%m-%d-%Y")

# Order the data frame by Year, Month, and Day
All <- All[order(year(All$Date), month(All$Date), day(All$Date)), ]

```

Join CAM + CSO

```{r}
All <- merge(All, CAM017, by = "Date", all = TRUE)
All <- All %>%
  rename(CAM017MG = MG)

```

```{r}
All <- na.omit(All)
```

### ALL CSOs

```{r}

ag_df_CR$Date <- ag_df_CR$date

ag_df_CR <- ag_df_CR[,-c(1) ]

#rm(CR2023, CR2022, CR2020, CR2021, CR2020, CR2019, CR2018, CR2017, CR2016, CR2015, df_CR)

ag_df_CR2 <- ag_df_CR


#Fix Date Formats of Future Joins
MWRACSOFinal$Date <- format(as.Date(MWRACSOFinal$Date, format=
                           "%m-%d-%Y"         ), "%m-%d-%Y")


ag_df_CR2$Date <- format(as.Date(ag_df_CR2$Date , format=
                           "%Y-%m-%d"          ), "%m-%d-%Y")


#join

All <- merge(MWRACSOFinal, ag_df_CR2, by = "Date", all = TRUE)

# Assuming your date column is named "Date"
All$Date <- as.Date(All$Date, format="%m-%d-%Y")

# Order the data frame by Year, Month, and Day
All <- All[order(year(All$Date), month(All$Date), day(All$Date)), ]

```

Join CAM + CSO

```{r}


CAM017 #Top CAM CSOs site 
All #ALL MWRA CSOs

All.CSOs <- merge(All, CAM017, by = "Date", all = TRUE)


```

###CSO OCCURRANCES

```{r}

# Assuming your data frame is named MWRACSOFinal and the column with CSO IDs is named CSO_ID
cso_counts <- as.data.frame(table(MWRACSOFinal$CSO_ID))

#011

```

#COMBINE ALL DATA

##MWRA X EPA

```{r}
ag_df_CR <- ag_df_CR %>%
    mutate(date = as.POSIXct(date, format = "%Y-%m-%d"))
ag_df_CR <- ag_df_CR %>% pad()

#All.CR_bacteria
ag_CR_bacteria <- ag_CR_bacteria %>% pad()
All.CR_bacteria <- left_join(ag_CR_bacteria, ag_df_CR,  by = c("Date" = "date"  ))

if (file.exists("All.CR_bacteria.rds")) {
  # If it's saved locally, load the data
 All.CR_bacteria <- readRDS("All.CR_bacteria.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
  saveRDS(All.CR_bacteria, "All.CR_bacteria.rds")
}
                            
#All.CR_secchi

ag_CR_secchi <- ag_CR_secchi %>% pad()
All.CR_secchi <- left_join(ag_CR_secchi, ag_df_CR,  by = c("Date" = "date"  ))


if (file.exists("All.CR_secchi.rds")) {
  # If it's saved locally, load the data
  CR2023 <- readRDS("All.secchi.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
  saveRDS(All.CR_secchi, "All.CR_secchi.rds")
}
                            

#All.CR.nutrients

ag_CR_nutrients <- ag_CR_nutrients %>% pad()
All.CR_nutrients <- left_join(ag_CR_nutrients, ag_df_CR,  by = c("Date" = "date"  ))

if (file.exists("All.CR_nutrients.rds")) {
  # If it's saved locally, load the data
  All.CR_nutrients <- readRDS("All.nutrients.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
  saveRDS(All.CR_nutrients, "All.CR_nutrients.rds")
}

                          
```

#LOGAN X EPA

```{r}

#TO BE FIXED 

#All.CR_bacteria
ag_Logan <- ag_Logan %>% pad()
All.ag_Logan <- left_join(ag_Logan, ag_df_CR,  by = c("DATE" = "date"  ))

if (file.exists("All.ag_Logan.rds")) {
  # If it's saved locally, load the data
  All.ag_Logan <- readRDS("All.ag_Logan.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
  saveRDS(All.ag_Logan, "All.ag_Logan.rds")
}

    
  

```

#CSOS

```{r}

CAM017 <- CAM017 %>% pad()
All.CAM017 <- left_join(CAM017, ag_df_CR,  by  = c("Date" = "date"  ))

if (file.exists("All.CAM017.rds")) {
  # If it's saved locally, load the data
  All.CAM017 <- readRDS("All.CAM017.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
  saveRDS(All.CAM017, "All.CAM017.rds")
} 
  
  
```

#Combine All Var

```{r}
All.Var1 <-  left_join(CAM017, ag_df_CR,  by  = c("Date" = "date"  ))

ag_Logan <- ag_Logan %>% pad()
All.Var1 <- left_join(ag_Logan, ag_df_CR,  by = c("DATE" = "date"  ))
All.Var2 <- left_join(All.Var1, CAM017,  by = c("DATE" = "Date"  ))
All.Var3 <- left_join(All.Var2, ag_CR_bacteria,  by = c("DATE" = "Date"  ))
All.Var4 <- left_join(All.Var3, ag_CR_nutrients,  by = c("DATE" = "Date"  ))
All.Var5 <- left_join(All.Var4, ag_CR_nutrients,  by = c("DATE" = "Date"  ))


All.Var <- All.Var5


if (file.exists("All.Var.rds")) {
  # If it's saved locally, load the data
  All.Var <- readRDS("All.Var.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
  saveRDS(All.Var, "All.Var.rds")
} 
  
  
```

#END
