==> 2_FIGURES/02_crossmap_EDM.Rmd <==

---
title: "02_crossmap_EDM"
output: html_document
date: "2024-04-03"
editor_options: 
  chunk_output_type: console
  markdown: 
    
    wrap: 72
---


#Library 

```{r}
library(rEDM)
library(tidyverse)
```

# "ag_df_CR" -> "DataFrame"

```{r}

# CONFIG: Define object name and paths
quoted_file_name <- "ag_df_CR"
rds_name <- paste0(quoted_file_name, ".rds")

local_path <- file.path(
  "C:/Users/mfman/OneDrive/Desktop/HAB2/HAB2/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling",
  rds_name
)

github_url <- paste0(
  "https://raw.githubusercontent.com/mfmanberg/HAB2/main/1_LIB/1_0_data_wrangling/1_0_2_data/1_0_2_2_data_wrangling/",
  rds_name
)


# Load from GitHub or local, if not save
if (!file.exists(local_path)) {
  tryCatch({
    con <- url(github_url)
    obj <- readRDS(con)
    close(con)  # Close the connection
    assign(quoted_file_name, obj)
    saveRDS(obj, local_path)
    message("Loaded from GitHub and saved locally.")
  }, error = function(e) {
    if (exists(quoted_file_name)) {
      saveRDS(get(quoted_file_name), local_path)
      message("GitHub failed. Saved in-memory object to local path.")
    } else {
      stop(paste0("Failed to load ", quoted_file_name, " from GitHub and no in-memory object exists to save."))
    }
  })
} else {
  assign(quoted_file_name, readRDS(local_path))
  message("Loaded from local path.")
}

# Clean up
rm(rds_name, quoted_file_name, local_path, github_url)



#DATAFRAME 


dataFrame <- ag_df_CR %>%
  mutate(yday = yday(date)) %>%
  select(yday, everything())

dataFrame <- dataFrame[, -c(2)]


# #cleanup 
# 
# rm(ag_df_CR)


```


#Surr functions

```{r}

#do_1_xmap -> RUN CCM. 

do_1_xmap <- function(dataFrame, E, tau, Tp, lib=NULL, pred=NULL, columns, target, sample, exclusionRadius, random = FALSE, showPlot = FALSE, ...) {

    if(is.null(lib)){
    lib <- paste("1",nrow(dataFrame))
    pred <- lib
  }

  # Run CCM
  ccm_result <- Simplex(
    dataFrame = dataFrame,
    E = E,
    lib = lib,
    pred = pred,
    tau = tau,
    Tp = Tp,
    columns = columns,
    target = target,
    exclusionRadius = exclusionRadius,
    ...
  )


  xmap_stats <- ComputeError(ccm_result$Observations,ccm_result$Predictions)


  return(xmap_stats)  # return mae, rmse, and rho
  # return(xmap_stats$rho)  # or return just rho
}

#CREATE SEASONAL SURROGATE, KERNEL SMOOTHING USING YDAY. SUBTRACT SEASONAL CYCLE TO GET RESIDUALS, SHUFFLE RESIDUALS. 

make_yday_surrogate <- function(yday,ts,bw=28,num_surr=1){
  
  
  if (any(!is.finite(ts))) {
    stop("SurrogateData(): input time series contained invalid values")
  }
  n = length(ts)
  
  seasonal_cyc = map_dbl(yday,\(yday_i) ksmooth(yday,ts,bandwidth=bw,x.points=yday_i)$y)
  seasonal_resid = ts - seasonal_cyc
  
  # randomize the temporal sequence of residuals and add them back to the cycle.
  
  return(sapply(1:num_surr, function(i) {
    seasonal_cyc + sample(seasonal_resid, n)
  }))
  
  
}

#REPLACE TARGET W/ SEASONAL SURROGATE

do_1_surr_xmap <- function(dataFrame, E, tau, Tp=0, columns, target, lib=NULL, pred=NULL, exclusionRadius, yday_col = "yday") {

  if(is.null(lib)){
    lib <- paste("1",nrow(dataFrame))
    pred <- lib
  }
  # Make a temporary data frame with a surrogate of `target`

  dataFrame_i <- dataFrame
  dataFrame_i[,target] <- make_yday_surrogate(dataFrame[,yday_col],dataFrame[,target])

  xmap_stats <- do_1_xmap(dataFrame_i, E=E, tau=tau, Tp=Tp, columns=columns, target=target,
                        lib=lib, pred=pred, exclusionRadius=exclusionRadius)


  return(xmap_stats)
}


```


#   N = 1 (TRUE),  ag_df_CR, phy

```{r}

# SET PARAMETERS

n_surr <- 1

tau_i <- 7 
E_i <- 3
Tp_i <- 0 

#PAR DATAFRAME VARS

df <- dataFrame
v_test_var <- c("ph","Celsius","spcond","do","turb","Salinity", "sin_season")
columns_i <- "phy"

#DYNAMIC PARAMS 

exclusionRadius_i <- ifelse(tau_i == 1, 1, tau_i - 1)

#CREATE SURROGATE DATAFRAME

results_surr_xmap <- map_dfr(v_test_var,\(target_i){
  
  result_i <- map_dfr(1:n_surr,\(index){
    temp <-  do_1_surr_xmap(
      dataFrame = df,
      E= E_i,
      tau=tau_i,
      Tp= Tp_i,
      columns= columns_i,
      target= target_i,
      exclusionRadius = exclusionRadius_i)
    
#OVERWRITE SURROGATE W/ TRUE
    
    
    temp <-  do_1_surr_xmap(
      dataFrame = df,
      E=E_i,
      tau=tau_i,
      Tp=Tp_i,
      columns=columns_i,
      target = columns_i, 
      exclusionRadius = exclusionRadius_i
      ) 
    temp$index = index
    return(temp)
    })
  
  
  #PULL RESULT
  
  result_i$target = target_i
  result_i <- result_i %>% select(target,index,everything())
  return(result_i)
  
  
})


#SAVE + CLEAN UP

n1_results_surr_xmap <- results_surr_xmap
rm(n_surr, tau_i, E_i, Tp_i, df, v_test_var, columns_i, exclusionRadius_i)


```

#   N = 500 (500 SURROGATES),  ag_df_CR, phy

```{r}

# SET PARAMETERS

n_surr <- 500

tau_i <- 7 
E_i <- 3
Tp_i <- 0 

#PAR DATAFRAME VARS

df <- dataFrame
v_test_var <- c("ph","Celsius","spcond","do","turb","Salinity", "sin_season")
columns_i <- "phy"

#DYNAMIC PARAMS 

exclusionRadius_i <- ifelse(tau_i == 1, 1, tau_i - 1)

#CREATE SURROGATE DATAFRAME

results_surr_xmap <- map_dfr(v_test_var,\(target_i){
  
  result_i <- map_dfr(1:n_surr,\(index){
    temp <-  do_1_surr_xmap(
      df,
      E= E_i,
      tau=tau_i,
      Tp= Tp_i,
      columns= columns_i,
      target= target_i,
      exclusionRadius = exclusionRadius_i)
    
#OVERWRITE SURROGATE W/ TRUE
    
   
    temp <-  do_1_surr_xmap(
      dataFrame = df,
      E=E_i,
      tau=tau_i,
      Tp=Tp_i,
      columns=columns_i,
      target = columns_i, 
      exclusionRadius = exclusionRadius_i
      ) 
    temp$index = index
    return(temp)
    })
  
  
  #PULL RESULT
  
  result_i$target = target_i
  result_i <- result_i %>% select(target,index,everything())
  return(result_i)
  
  
})

#SAVE + CLEAN UP

n500_results_surr_xmap <- results_surr_xmap
rm(n_surr, tau_i, E_i, Tp_i, df, v_test_var, columns_i, exclusionRadius_i)


```

#LOAD IF NOT SAVE 

```{r}
n1_results_surr_xmap <- results_surr_xmap

n500_results_surr_xmap <- results_surr_xmap2

saveRDS(n500_results_surr_xmap, file = "n500_results_surr_xmap.rds" )

results_surr_xmap2 <- results_surr_xmap
```


#Surr Test Table 
```{r}




surrtable <- n1_results_surr_xmap %>%
  rowwise() %>%
  mutate(
    Psurr = mean(
      n500_results_surr_xmap$rho[n500_results_surr_xmap$target == target] >= rho, #mean (psurr >= pmax)
      na.rm = TRUE
    ) 
  ) %>%
  ungroup()  %>%
  select(-index, -MAE, -RMSE)



# Compute correlation between rho and Psurr
rho_psurr_corr <- cor(surrtable$rho, surrtable$Psurr, use = "complete.obs")

# Add the correlation as a new column in every row
surrtable <- surrtable %>%
  mutate(correlation = rho_psurr_corr)





# First, join the two datasets to get matching rho values
joined_data <- n1_results_surr_xmap %>%
  select(target, rho_n1 = rho) %>%
  inner_join(n500_results_surr_xmap %>% select(target, rho_n500 = rho), by = "target")

# Calculate correlation for each target
correlations <- joined_data %>%
  group_by(target) %>%
  summarise(rho_correlation = cor(rho_n1, rho_n500))

# Merge correlation into the original surrtable
surrtable <- surrtable %>%
  left_join(correlations, by = "target")





n500_results_surr_xmap <- n500_results_surr_xmap %>%
  group_by(target) %>%
  filter(rho == max(rho, na.rm = TRUE))  %>%
  select(-index, -MAE, -RMSE)


surrtable2 <- left_join(n500_results_surr_xmap,n1_results_surr_xmap , by = "target")  %>%
  select(-index, -MAE, -RMSE) %>%
  rename(
    pmax = rho.x, #n1
    psurr = rho.y #n500
)

surrtable2 <- surrtable2 %>%
  left_join(select(surrtable, target, Pvalue), by = "target")























surrccm_phyph_n500 <- results_surr_xmap

if (file.exists("surrccm_phyph_n500.rds")) {
  # If it's saved locally, load the data
  surrccm_phyph_n500 <- readRDS("surrccm_phyph_n500.rds")
} else {
  # If it's not saved locally, read it from Google Sheets and save it locally
surrccm_phyph_n500 <- surrccm_phyph_n500
  saveRDS(surrccm_phyph_n500, "surrccm_phyph_n500.rds")
}

library(dplyr)


avg_surrccm_phyph_n500 <- surrccm_phyph_n500 %>%
  group_by(target) %>%
  summarise(
    mean_MAE = mean(MAE, na.rm = TRUE),
    mean_rho = mean(rho, na.rm = TRUE),
    mean_RMSE = mean(RMSE, na.rm = TRUE)
  )


library(gridExtra)
library(grid)

# Define output file path
output_file <- "avg_surrccm_phyph_n500.png"

# Open PNG graphics device
png(output_file, width = 1200, height = 600, res = 150)  # Adjust size & resolution

# Create table and plot
grid.newpage()
grid.table(output_file)

# Close the graphics device
dev.off()


# surrccm_phyph_n500 <- surrccm_phyph_n500 %>%
#   group_by(target) %>%
#   filter(rho == max(rho, na.rm = TRUE))  %>%
#   select(-index, -MAE, -RMSE)


# Create the new column in df of the REAL covariate xmaps
# surrtable <- surrccm_phyph_n1 %>% # change surrccm_phyph_n1 so it's the actual xmap values not 1 surrogate

surrtable <- results_xmap %>%
  rowwise() %>%
  mutate(
    Pvalue = mean(
      surrccm_phyph_n500$rho[surrccm_phyph_n500$target == target] >= rho, #mean (psurr >= pmax)
      na.rm = TRUE
    )
  ) %>%
  ungroup()


surrccm_phyph_n500 <- surrccm_phyph_n500 %>%
  group_by(target) %>%
  filter(rho == max(rho, na.rm = TRUE))  %>%
  select(-index, -MAE, -RMSE)


surrtable2 <- left_join(surrccm_phyph_n500,surrccm_phyph_n1 , by = "target")  %>%
  select(-index, -MAE, -RMSE) %>%
  rename(
    pmax = rho.x, #n1
    psurr = rho.y #n500
)

surrtable2 <- surrtable2 %>%
  left_join(select(surrtable, target, Pvalue), by = "target")


```



#CMAPS
go off univar tau
- help us pick timescale of analysis appropiate for focal
rather than refight value 

df 
E_i
tau_i
columns
target



```{r}

library(rEDM)

tau_ccm <- 7

E_ccm <- 3


critical.r <- function( n, alpha = .05 ) {
   df <- n - 2
  critical.t <- qt(alpha/2, df, lower.tail = F)
  critical.r <- sqrt( (critical.t^2) / ( (critical.t^2) + df ) )
   return(critical.r)
}




#ph

cmap.ph.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "ph", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1200 1300 1339", sample = 100, 
                  # exclusionRadius = tau_ccm, 
                   random = TRUE, showPlot = TRUE) 

#CELSIUS

cmap.Celsius.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "Celsius", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", sample = 100, 
                        #exclusionRadius = tau_ccm, 
                        random = TRUE, showPlot = TRUE) 


#Celsius 
cmap.Celsius.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "Celsius", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", sample = 100,
                        #exclusionRadius = tau_ccm,
                        random = TRUE, showPlot = TRUE) 



#spcond

cmap.spcond.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "spcond", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", sample = 100, 
                    #   exclusionRadius = tau_ccm, 
                       random = TRUE, showPlot = TRUE) 



#do 
cmap.do.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "do", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", 
                   sample = 100,
                 #  exclusionRadius = tau_ccm, 
                   random = TRUE, showPlot = TRUE) 

#TURB

cmap.turb.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "turb", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", sample = 100, 
                    # exclusionRadius = tau_ccm, 
                     random = TRUE, showPlot = TRUE) 


#SALINITY

cmap.Salinity.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "Salinity", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", sample = 100,
                       #  exclusionRadius = tau_ccm,
                         random = TRUE, showPlot = TRUE) 

cmap.Salinity.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-tau_ccm, Tp = 0, columns = "Salinity", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", sample = 100,
                      #   exclusionRadius = tau_ccm, 
                         random = TRUE, showPlot = TRUE) 

#sin_season

cmap.sin_season.phy <- CCM(dataFrame = ag_df_CR, E = E_ccm, tau=-7, Tp = 0, columns = "sin_season", target = "phy", libSizes = "20 40 60 80 100 140 180 240 300 400 500 600 700 800 900 1000 1100 1339", sample = 100, 
                          # exclusionRadius = tau_i, 
                           random = TRUE, showPlot = TRUE) 


```


#CCM Helper Function - Plot


```{r}

#Random = TRUE


tau_i <- 7

#ag_df_CR <- readRDS("C:\\Users\\mfman\\OneDrive\\Desktop\\HAB2\\HAB2\\1)LIB\\ag_df_CR.rds")
ag_df_CR <- na.omit(ag_df_CR)

library(rEDM)


plot_ccm_result <- function(cmap_result, var_name, target_name = "Phycocanin (RFU)", corr_value, critical_corr, 
                            ylim_vals = c(0, 1), legend_labels = NULL, legend_colors = NULL, legend_lty = NULL) {
  # Custom tick setup
  lib_ticks <- c(20, 40, 60, 80, 100, 200, 400, 600, 800, 1000, 1100, 1200, 1300, 1339)
  sqrt_ticks <- sqrt(lib_ticks)

  # Prepare x
  sqrt_x <- sqrt(cmap_result[, 1])

  # Plot
  plot(x = sqrt_x, y = cmap_result[, 3], type = "l",
       main = paste(var_name, "V.", target_name),
       xlab = "LibSize", ylab = "Rho", col = "blue",
       ylim = ylim_vals, xaxt = "n")
  
  lines(x = sqrt_x, y = cmap_result[, 3], col = "red")
  lines(x = sqrt_x, y = cmap_result[, 2], col = "blue")

  abline(h = abs(corr_value), col = "green", lty = 2)
  abline(h = critical_corr, col = "darkgreen", lty = 2)

  axis(1, at = sqrt_ticks, labels = lib_ticks)

  # If legend arguments are provided, add legend
  if (!is.null(legend_labels)) {
    legend("topright", 
           legend = legend_labels, 
           col = legend_colors, 
           lty = legend_lty)
  }
}


#Pull Corr function 

pull_corr <- function(data, predictor) {
  # Create the formula
  formula <- as.formula(paste("phy ~", predictor))
  
  # Fit the linear model
  model <- lm(formula, data = data)
  
  # Extract R-squared from the summary
  r_squared <- summary(model)$r.squared
  
  # Calculate correlation (R) from R-squared
  r <- sqrt(r_squared)
  
  # If the slope is negative, make R negative
  if (coef(model)[2] < 0) {
    r <- -r
  }
  
  return(r)
}



# Call the helper function

#pH

plot_ccm_result(cmap.ph.phy, "ph", corr_value = pull_corr(ag_df_CR, "ph"), critical_corr = critical.r(1374), 
                legend_labels = c("ph:Phycocanin", "Phycocanin:ph", "Correlation", "Critical Correlation"),
                legend_colors = c("blue", "red", "green", "darkgreen"),
                legend_lty = c(1, 1, 2, 2))
legend("topright", 
       inset = c(-0.2, 0), 
       xpd = TRUE, 
       legend = legend_labels,
       col = legend_colors, 
       lty = legend_lty,
       bty = "n")

#TEMPERATURE


plot_ccm_result(cmap.Celsius.phy, "Temperature (°C)", corr_value = pull_corr(ag_df_CR, "Celsius"), critical_corr = critical.r(1374), 
                legend_labels = c("Celsius:Phycocanin", "Phycocanin:Celsius", "Correlation", "Critical Correlation"),
                legend_colors = c("blue", "red", "green", "darkgreen"),
                legend_lty = c(1, 1, 2, 2))
legend("topright", 
       inset = c(-0.2, 0), 
       xpd = TRUE, 
       legend = legend_labels,
       col = legend_colors, 
       lty = legend_lty,
       bty = "n")
                                                                                     
                                                                                                                     
# SPECIFIC CONDUCTIVITY  V. Phyocanin (RFU)

plot_ccm_result(cmap.spcond.phy, "Specific Conductivity (mS/cm)", 
                corr_value = pull_corr(ag_df_CR, "spcond"), 
                critical_corr = critical.r(1374),
                legend_labels = c("Specific Conductivity:Phycocanin", "Phycocanin:Specific Conductivity", "Correlation", "Critical Correlation"),
                legend_colors = c("blue", "red", "green", "darkgreen"),
                legend_lty = c(1, 1, 2, 2))
legend("topright", 
       inset = c(-0.2, 0), 
       xpd = TRUE, 
       legend = legend_labels,
       col = legend_colors, 
       lty = legend_lty,
       bty = "n")

#DISSOLVED OXYGEN

plot_ccm_result(cmap.do.phy, "Dissolved Oxygen (mg/L)", 
                corr_value = pull_corr(ag_df_CR, "do"), 
                critical_corr = critical.r(1374),
                legend_labels = c("Dissolved Oxygen:Phycocanin", "Phycocanin:Dissolved Oxygen", "Correlation", "Critical Correlation"),
                legend_colors = c("blue", "red", "green", "darkgreen"),
                legend_lty = c(1, 1, 2, 2))
legend("topright", 
       inset = c(-0.2, 0), 
       xpd = TRUE, 
       legend = legend_labels,
       col = legend_colors, 
       lty = legend_lty,
       bty = "n")


#TURBIDITY
plot_ccm_result(cmap.turb.phy, "Turbidity (FNU)", 
                corr_value = pull_corr(ag_df_CR, "turb"), 
                critical_corr = critical.r(1374),
                legend_labels = c("Turbidity:Phycocanin", "Phycocanin:Turbidity", "Correlation", "Critical Correlation"),
                legend_colors = c("blue", "red", "green", "darkgreen"),
                legend_lty = c(1, 1, 2, 2))
legend("topright", 
       inset = c(-0.2, 0), 
       xpd = TRUE, 
       legend = legend_labels,
       col = legend_colors, 
       lty = legend_lty,
       bty = "n")

# SALINITY
plot_ccm_result(cmap.Salinity.phy, "Practical Salinity (S/m)", 
                corr_value = pull_corr(ag_df_CR, "Salinity"),
                critical_corr = critical.r(1374),
                legend_labels = c("Salinity:Phycocanin", "Phycocanin:Salinity", "Correlation", "Critical Correlation"),
                legend_colors = c("blue", "red", "green", "darkgreen"),
                legend_lty = c(1, 1, 2, 2))
legend("topright", 
       inset = c(-0.2, 0), 
       xpd = TRUE, 
       legend = legend_labels,
       col = legend_colors, 
       lty = legend_lty,
       bty = "n")

#SIN_SEASON 

plot_ccm_result(cmap.sin_season.phy, "Seasonality (Days)", 
                corr_value = pull_corr(ag_df_CR, "sin_season"), 
                critical_corr = critical.r(1374),
                legend_labels = c("Seasonality:Phycocanin", "Phycocanin:Seasonality", "Correlation", "Critical Correlation"),
                legend_colors = c("blue", "red", "green", "darkgreen"),
                legend_lty = c(1, 1, 2, 2))
legend("topright", 
       inset = c(-0.2, 0), 
       xpd = TRUE, 
       legend = legend_labels,
       col = legend_colors, 
       lty = legend_lty,
       bty = "n")

```
