==> 2_FIGURES/03_multivariate_EDM.Rmd <==

---
title: "03_multivariate_EDM"
output: html_document
date: "2024-04-03"
editor_options: 
  chunk_output_type: console
  markdown: 
    wrap: 72
---


# EDM Mulit-variate Analysis 


#Library
```{r}

library(rEDM)
library(tidyverse)

```



# Multivariate 

## KITCHEN SINK 7 DAYS/WEEKLY

#### Data Config 7 DAYS KITCHEN SINK

```{r}

tau_i <- 7 #Days out

library(rEDM)

#UniVariate

ag_df_CR <- na.omit(ag_df_CR)
ag_df_CR$date <- as.Date(ag_df_CR$date)


first_row_2023 <- which(format(ag_df_CR$date, "%Y") == "2023")[1]

print(first_row_2023)

lib_ins <- c(1,1198) # data up through end of 2022; this is the training set
lib_oos <- c(1199,1374) # Pred

columns_i <- c( "chl" ,     "phy"   ,   "ph"   ,    "Celsius" , "spcond" ,  "do" ,      "turb"  ,   "Salinity" , "sin_season")



```


#### Simplex - Find Optimal E 7 DAYS KITCHEN SINK
```{r}

#SELECT E - SIMPLEX

simplex_selectE <- EmbedDimension(dataFrame = ag_df_CR, columns = columns_i, target = "phy",
                        tau=-tau_i,Tp=tau_i, maxE = 25, 
lib = lib_ins, pred = lib_oos, showPlot = TRUE) #lib train, #pred is pred 


best_simplex_selectE_row <- simplex_selectE[which.max(simplex_selectE$rho), ]
simplexE <- best_simplex_selectE_row$E


max_simplex_phy_tau7 <-best_simplex_selectE_row$rho


KS_max_simplex_phy_tau7 <- max_simplex_phy_tau7
 

```

#### SMap - Select Theta No Ex 7 DAYS KITCHEN SINK

```{r}

#SELECT THETA

#w/o exclusion radius

simplex_pred <- PredictNonlinear(dataFrame = ag_df_CR, columns = columns_i, target = "phy",
                                    E= simplexE ,
                                    tau=-tau_i,Tp=tau_i,#exclusionRadius = tau_i,
                                    lib = lib_ins, pred = lib_oos, showPlot = TRUE, embedded = TRUE) 


best_simplex_theta_row <- print(simplex_pred[which.max(simplex_pred$rho), ])


simplextheta <- best_simplex_theta_row$Theta

```


#### SMap - Eval Exclusion Radius 7 DAYS  KITCHEN SINK - TO DO


HELPER FUNCTION #ADD # here


eval_ex_radius <- function(dataFrame, columns, target, E, tau_i, lib_ins, lib_oos) {
  
  # Create an empty list to store results
  results_list <- list()
  
  # Loop over exclusionRadius values from 1 to 10
  for (ex_radius in 1:10) { #choose ex values
    
    tryCatch({
      result <- PredictNonlinear(
        dataFrame = dataFrame,
        columns = columns,
        target = target,
        E = E,
        tau = -tau_i,  
        Tp = tau_i,
        exclusionRadius = ex_radius,
        lib = lib_ins,
        pred = lib_oos,
        showPlot = FALSE
      )
      
      # Add exclusionRadius info
      result$exclusionRadius <- ex_radius
      
      # Store result
      results_list[[ex_radius]] <- result
      
    }, error = function(e) {
      # If error happens, store NA and print warning
      message(sprintf("Error at exclusionRadius = %d: %s", ex_radius, e$message))
      results_list[[ex_radius]] <- NA
    })
    
  }
  
  # Remove any NA results
  results_list_clean <- results_list[!sapply(results_list, is.na)]
  
  # Combine into a single dataframe
  if (length(results_list_clean) > 0) {
    results_select_ex <- do.call(rbind, results_list_clean)
  } else {
    results_select_ex <- data.frame()
  }
  
  return(results_select_ex)
}

eval_ex_radius_results <- eval_ex_radius(
  dataFrame = ag_df_CR,
  columns = "phy",
  target = "phy",
  E = 5,
  tau_i = tau_i,
  lib_ins = lib_ins,
  lib_oos = lib_oos
)

simplexE_ex <- print(eval_ex_radius_results[which.max(eval_ex_radius_results$rho), ])

simplexE_ex <- simplexE_ex$rho


#### Smap - Make Predictions 7 DAYS KITCHEN SINK


```{r}

#PREDICT - SMAP

smap_pred <- SMap(dataFrame = ag_df_CR, columns = columns_i, target = "phy",
lib = lib_ins, pred = lib_oos, E=simplexE,theta= simplextheta, tau = -tau_i, Tp = tau_i,
embedded = TRUE)

 smap_results <- as.data.frame(smap_pred[["predictions"]])


smapstats <- ComputeError(
                          smap_results$Observations,
                       smap_results$Predictions)

#RESULT: 

max_smap_phy_tau7 <- print(smapstats$rho)

KS_max_smap_phy_tau7 <- max_smap_phy_tau7


```




# Multivariate 

## MULTI 7 DAYS/WEEKLY

#### Data Config 7 DAYS MULTI

```{r}

tau_i <- 7 #Days out

library(rEDM)

#UniVariate

ag_df_CR <- na.omit(ag_df_CR)
ag_df_CR$date <- as.Date(ag_df_CR$date)


first_row_2023 <- which(format(ag_df_CR$date, "%Y") == "2023")[1]

print(first_row_2023)

lib_ins <- c(1,1198) # data up through end of 2022; this is the training set
lib_oos <- c(1199,1374) # Pred

columns_i <- c("Celsius" , "spcond" , "turb")



```


#### Simplex - Find Optimal E 7 DAYS MULTI
```{r}

#SELECT E - SIMPLEX

simplex_selectE <- EmbedDimension(dataFrame = ag_df_CR, columns = columns_i, target = "phy",
                        tau=-tau_i,Tp=tau_i, maxE = 50, 
lib = lib_ins, pred = lib_oos, showPlot = TRUE) #lib train, #pred is pred 


best_simplex_selectE_row <- simplex_selectE[which.max(simplex_selectE$rho), ]
simplexE <- best_simplex_selectE_row$E


max_simplex_phy_tau7 <-best_simplex_selectE_row$rho


multi_max_simplex_phy_tau7 <- max_simplex_phy_tau7
 

```

#### SMap - Select Theta No Ex 7 DAYS MULTI

```{r}

#SELECT THETA

#w/o exclusion radius

simplex_pred <- PredictNonlinear(dataFrame = ag_df_CR, columns = columns_i, target = "phy",
                                    E= simplexE ,
                                    tau=-tau_i,Tp=tau_i,#exclusionRadius = tau_i,
                                    lib = lib_ins, pred = lib_oos, showPlot = TRUE, embedded = TRUE) 


best_simplex_theta_row <- print(simplex_pred[which.max(simplex_pred$rho), ])


simplextheta <- best_simplex_theta_row$Theta

```


#### SMap - Eval Exclusion Radius 7 DAYS  MULTI - TO DO


HELPER FUNCTION #ADD # here


eval_ex_radius <- function(dataFrame, columns, target, E, tau_i, lib_ins, lib_oos) {
  
  # Create an empty list to store results
  results_list <- list()
  
  # Loop over exclusionRadius values from 1 to 10
  for (ex_radius in 1:10) { #choose ex values
    
    tryCatch({
      result <- PredictNonlinear(
        dataFrame = dataFrame,
        columns = columns,
        target = target,
        E = E,
        tau = -tau_i,  
        Tp = tau_i,
        exclusionRadius = ex_radius,
        lib = lib_ins,
        pred = lib_oos,
        showPlot = FALSE
      )
      
      # Add exclusionRadius info
      result$exclusionRadius <- ex_radius
      
      # Store result
      results_list[[ex_radius]] <- result
      
    }, error = function(e) {
      # If error happens, store NA and print warning
      message(sprintf("Error at exclusionRadius = %d: %s", ex_radius, e$message))
      results_list[[ex_radius]] <- NA
    })
    
  }
  
  # Remove any NA results
  results_list_clean <- results_list[!sapply(results_list, is.na)]
  
  # Combine into a single dataframe
  if (length(results_list_clean) > 0) {
    results_select_ex <- do.call(rbind, results_list_clean)
  } else {
    results_select_ex <- data.frame()
  }
  
  return(results_select_ex)
}

eval_ex_radius_results <- eval_ex_radius(
  dataFrame = ag_df_CR,
  columns = "phy",
  target = "phy",
  E = 5,
  tau_i = tau_i,
  lib_ins = lib_ins,
  lib_oos = lib_oos
)

simplexE_ex <- print(eval_ex_radius_results[which.max(eval_ex_radius_results$rho), ])

simplexE_ex <- simplexE_ex$rho


#### Smap - Make Predictions 7 DAYS MULTI


```{r}

#PREDICT - SMAP

smap_pred <- SMap(dataFrame = ag_df_CR, columns = columns_i, target = "phy",
lib = lib_ins, pred = lib_oos, E=simplexE,theta= simplextheta, tau = -tau_i, Tp = tau_i,
embedded = TRUE)

 smap_results <- as.data.frame(smap_pred[["predictions"]])


smapstats <- ComputeError(
                          smap_results$Observations,
                       smap_results$Predictions)

#RESULT: 

max_smap_phy_tau7 <- print(smapstats$rho)

multi_max_smap_phy_tau7 <- max_smap_phy_tau7


```



Next Steps

Convergent Cross Mapping All possible Covariates in one data frame as
different columns other variables all the endogenous bigoeochemical and
human drivers management driver and - rainfall and CSOs know now half
week to weekly time scale we see has signature until we have the
corrected values to keep the diff 9 to noon 4 days negative auto, use
linear s-map To make forecasts: predict 1st differences and add back
into data prediction for tomorrow is value for tomorrow + value for
today

list of measurments and list of weekly measurments look at Tahoe +
geneva

plot minutes vs. volume CSOs - rainfall causal associations? analysis:
same trend = synchornity not = one is low and flat is oneway causal\
number at end matters, consider as correlation coeff, rho ccm at max
library size

look at full matrix with nutrient time series later on

2 different data matrixes 1. time-scale 2. timespan

#Multiview


ccm uses y to predict x, if true, x can affect y ccm wraps (Simplex
target = x, columns = y)

multivariate Simplex ( Embedded = True, columns = c("x1, "x2)) set CCM
samples to 10 to see result

Here, we demonstrate this idea using the Multiview() function with the
3-species data used above. Multiview() operates by constructing all
possible embeddings of dimension E with lag up to E-1. These embeddings
are ranked by forecast skill (rho) over the lib portion of the data. The
individual forecasts for the top multiview embeddings are then averaged
together. If multiview is not specified it is set to sqrt(C) where C is
the number of E-dimensional combinations created from all data vectors.

View lists the various combinations of data embedding vectors used for
the forecasts along with their prediction statistics. Predictions
returns the final averaged multiview projections.

E is amounts of lags. univar time lags D is dimension of individual
multivariate models, number of vars we're taking at a time. axis to
embed. Additional Analysis fills in mechnaistically. Can set lib = pred
At least E of 2, after E \> 5 com sci problem



#```{r}


#Mutltiview

#Omit NAs 

ag_df_CR <- na.omit(ag_df_CR)


# Find the first row number for the first date in 2023
first_row_2023 <- which(format(ag_df_CR$date, "%Y") == "2023")[1]

print(first_row_2023)

#Check class

ag_df_CR$date <- as.numeric(ag_df_CR$date)
library(rEDM)


#**E3 D5 IN SAMPLE**

# Check if the file exists
if(file.exists("Multi_E3_D5.rds")) {
  # Load the file if it exists
  Multi_E3_D5 <- readRDS("Mutli_E3D5.rds")
} else {
  # If the file does not exist, create Multi_E2_D3
  Multi_E3_D5 <- Multiview(dataFrame = ag_df_CR, lib = "1 1014", pred = "1 1014", E =3, D=5,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")
  # Save Multi_E2_D3 to file
  saveRDS(Multi_E3_D5, "Mutli_E2D5.rds")
}




   Multi_E3_D5_obs <- Multiview(dataFrame = ag_df_CR, lib = "1 1014", pred = "1 1014", E =3, D=5,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")
   
     Multi_E3_D5_pred <- Multiview(dataFrame = ag_df_CR, lib = "1 1014", pred = "1 1189", E =3, D=5,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")
                                

     
     
#**E2 D3 IN SAMPLE**

Multi_E2_D3 <- "Multi_E2_D3.rds"
if(!file.exists("Multi_E2_D3.rds")){
 Multi_E2_D3 <- Multiview(dataFrame = ag_df_CR, lib = "1 1014", pred = "1 1014", E = 2, D=3,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")
    save(Multi_E2_D3, file = "Multi_E2_D3.rds")
} else{
  load(Multi_E2_D3)
}


#HEAD


Multi_E3_D5_obs <- Multiview(dataFrame = ag_df_CR, lib = "1 1014", pred = "1 1014", E =3, D=5,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")


Multi_E3_D5_pred <- Multiview(dataFrame = ag_df_CR, lib = "1 1014", pred = "1 1189", E =3, D=5,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")


saveRDS(Multi_E3_D5_obs, file = "Multi_E3_D5_obs.rds")
saveRDS(Multi_E3_D5_pred, file = "Multi_E3_D5_pred.rds")

```

```{r}

#post 2016 fix
library(rEDM)
library(tidyverse)

  ag_df_CR$date <- as.Date(ag_df_CR$date)
  
Multi_E3_D5_pred <- Multiview(dataFrame = ag_df_CR, lib = "1 1198
", pred = "1 1339", E =3, D=5,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")

saveRDS(Multi_E3_D5_pred, file = "Multi_E3_D5_pred.rds")


 
Multi_E2_D3_pred <- Multiview(dataFrame = ag_df_CR, lib = "1 1198
", pred = "1 1339", E =2, D=3,
                            columns = "chl phy ph Celsius spcond do turb Salinity sin_season", target = "phy")

saveRDS(Multi_E2_D3_pred, file = "Multi_E2_D3_pred.rds")

#```



## UNI 7 DAYS/WEEKLY For Comparision

#### Data Config 7 DAYS

```{r}

tau_i <- 7 #Days out

library(rEDM)

#UniVariate

ag_df_CR <- na.omit(ag_df_CR)
ag_df_CR$date <- as.Date(ag_df_CR$date)


first_row_2023 <- which(format(ag_df_CR$date, "%Y") == "2023")[1]

print(first_row_2023)

lib_ins <- c(1,1198) # data up through end of 2022; this is the training set
lib_oos <- c(1199,1374) # Pred


```


#### Simplex - Find Optimal E 7 DAYS
```{r}

#SELECT E - SIMPLEX

simplex_selectE <- EmbedDimension(dataFrame = ag_df_CR, columns = "phy", target = "phy",
                        tau=-tau_i,Tp=tau_i, maxE = 20, 
lib = lib_ins, pred = lib_oos, showPlot = TRUE) #lib train, #pred is pred 



best_simplex_selectE_row <- simplex_selectE[which.max(simplex_selectE$rho), ]
simplexE <- best_simplex_selectE_row$E


max_simplex_phy_tau7 <-best_simplex_selectE_row$rho

max_simplex_phy_tau7

```

#### SMap - Select Theta No Ex 7 DAYS

```{r}

#SELECT THETA

#w/o exclusion radius

simplex_pred <- PredictNonlinear(dataFrame = ag_df_CR, columns = "phy", target = "phy",
                                    E= simplexE ,
                                    tau=-tau_i,Tp=tau_i,#exclusionRadius = tau_i,
                                    lib = lib_ins, pred = lib_oos, showPlot = TRUE) 


best_simplex_theta_row <- print(simplex_pred[which.max(simplex_pred$rho), ])


simplextheta <- best_simplex_theta_row$Theta

#RESULT: 



```


#### SMap - Eval Exclusion Radius 7 DAYS - TO DO


HELPER FUNCTION #ADD # here


eval_ex_radius <- function(dataFrame, columns, target, E, tau_i, lib_ins, lib_oos) {
  
  # Create an empty list to store results
  results_list <- list()
  
  # Loop over exclusionRadius values from 1 to 10
  for (ex_radius in 1:10) { #choose ex values
    
    tryCatch({
      result <- PredictNonlinear(
        dataFrame = dataFrame,
        columns = columns,
        target = target,
        E = E,
        tau = -tau_i,  
        Tp = tau_i,
        exclusionRadius = ex_radius,
        lib = lib_ins,
        pred = lib_oos,
        showPlot = FALSE
      )
      
      # Add exclusionRadius info
      result$exclusionRadius <- ex_radius
      
      # Store result
      results_list[[ex_radius]] <- result
      
    }, error = function(e) {
      # If error happens, store NA and print warning
      message(sprintf("Error at exclusionRadius = %d: %s", ex_radius, e$message))
      results_list[[ex_radius]] <- NA
    })
    
  }
  
  # Remove any NA results
  results_list_clean <- results_list[!sapply(results_list, is.na)]
  
  # Combine into a single dataframe
  if (length(results_list_clean) > 0) {
    results_select_ex <- do.call(rbind, results_list_clean)
  } else {
    results_select_ex <- data.frame()
  }
  
  return(results_select_ex)
}

eval_ex_radius_results <- eval_ex_radius(
  dataFrame = ag_df_CR,
  columns = "phy",
  target = "phy",
  E = 5,
  tau_i = tau_i,
  lib_ins = lib_ins,
  lib_oos = lib_oos
)

simplexE_ex <- print(eval_ex_radius_results[which.max(eval_ex_radius_results$rho), ])

simplexE_ex <- simplexE_ex$rho


#### Smap - Make Predictions 7 DAYS


```{r}

#PREDICT - SMAP

smap_pred <- SMap(dataFrame = ag_df_CR, columns = "phy", target = "phy",
lib = lib_ins, pred = lib_oos, E=simplexE,theta= simplextheta, tau = -tau_i, Tp = tau_i)

 smap_results <- as.data.frame(smap_pred[["predictions"]])


smapstats <- ComputeError(
                          smap_results$Observations,
                       smap_results$Predictions)

#RESULT: 

max_smap_phy_tau7 <- print(smapstats$rho)

```

```{r}

df_Multi_E2_D3_pred <- data.frame(Multi_E2_D3_pred$View)

library(gridExtra)
library(grid)

png("df_Multi_E2_D3_pred.png", width = 1000, height = 700)
grid.table(df_Multi_E2_D3_pred)
dev.off()


```




# Multivariate Hist 

```{r}

library(tibble)
library(ggplot2)


Multi_E2_D3_rho <- ComputeError(
  unlist(Multi_E2_D3_pred$Predictions$Observations),
  unlist(Multi_E2_D3_pred$Predictions$Predictions)
)
Multi_E2_D3_rho <- Multi_E2_D3_rho$rho 

# 
# Multi_E3_D5_rho <- ComputeError(
#   unlist(Multi_E3_D5_pred$Predictions$Observations),
#   unlist(Multi_E3_D5_pred$Predictions )
# )
# Multi_E3_D5_rho <- Multi_E3_D5_rho$rho 


# Model names
model_names <- c(
  "Univariate EDM 7 Day Simplex", "Univariate EDM 7 Day S-map",
  "Multivaraite 'Kitchen Sink' 7 Day Simplex", "Multivaraite 'Kitchen Sink' 7 Day S-map",
  "Multivaraite EDM 7 Day Simplex", "Multivariate EDM 7 Day S-map",
  "Multiview: E = 2, D = 3"
)

# Corresponding broad categories
model_type <- c(
  "Univariate", "Univariate",
  "Multivariate 'Kitchen Sink'", "Multivariate 'Kitchen Sink'",
  "Multivariate", "Multivariate",
  "Multiview"
)

# Create the tibble
multi_daily_rho <- tibble(
  name = factor(model_names, levels = model_names),
  value = multi_rho_df$value,
  model_type = factor(model_type, levels = c("Univariate", "Multivariate 'Kitchen Sink'", "Multivariate", "Multiview"))
)

# Plot
ggplot(multi_daily_rho, aes(x = name, y = value, fill = model_type)) +
  geom_col() +
  scale_fill_manual(values = c( 
    "Univariate" = "orange", 
    "Multivariate 'Kitchen Sink'" = "darkorange",
    "Multivariate" = "red",
    "Multiview" = "darkred")) +
  theme_minimal() +
  labs(
    x = "Model", 
    y = "Prediction Skill (ρ)", 
    title = "Phycocanin (RFU) Multivariate & Univariate EDM Model Comparison", 
    fill = "Model Type"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1),
    legend.position = "top"
  )

```

#Option A:  

#Option B:  time series of predicted and observed





#END



## UNI 7 DAYS/WEEKLY CHL

#### Data Config 7 DAYS CHL

```{r}

tau_i <- 7 #Days out

library(rEDM)

#UniVariate

ag_df_CR <- na.omit(ag_df_CR)
ag_df_CR$date <- as.Date(ag_df_CR$date)


first_row_2023 <- which(format(ag_df_CR$date, "%Y") == "2023")[1]

print(first_row_2023)

lib_ins <- c(1,1198) # data up through end of 2022; this is the training set
lib_oos <- c(1199,1374) # Pred


```


#### Simplex - Find Optimal E 7 DAYS CHL 
```{r}

#SELECT E - SIMPLEX

simplex_selectE <- EmbedDimension(dataFrame = ag_df_CR, columns = "phy", target = "chl",
                        tau=-tau_i,Tp=tau_i, maxE = 20, 
lib = lib_ins, pred = lib_oos, showPlot = TRUE) #lib train, #pred is pred 



best_simplex_selectE_row <- simplex_selectE[which.max(simplex_selectE$rho), ]
simplexE <- best_simplex_selectE_row$E


max_simplex_phy_tau7 <-best_simplex_selectE_row$rho

chl_max_simplex_phy_tau7 <- max_simplex_phy_tau7

chl_simplex_selectE_tau7 <- simplex_selectE

```

#### SMap - Select Theta No Ex 7 DAYS CHL

```{r}

#SELECT THETA

#w/o exclusion radius

simplex_pred <- PredictNonlinear(dataFrame = ag_df_CR, columns = "phy", target = "chl",
                                    E= simplexE ,
                                    tau=-tau_i,Tp=tau_i,#exclusionRadius = tau_i,
                                    lib = lib_ins, pred = lib_oos, showPlot = TRUE) 


best_simplex_theta_row <- print(simplex_pred[which.max(simplex_pred$rho), ])


simplextheta <- best_simplex_theta_row$Theta

#RESULT: 



```


#### SMap - Eval Exclusion Radius 7 DAYS PHY - TO DO


HELPER FUNCTION #ADD # here


eval_ex_radius <- function(dataFrame, columns, target, E, tau_i, lib_ins, lib_oos) {
  
  # Create an empty list to store results
  results_list <- list()
  
  # Loop over exclusionRadius values from 1 to 10
  for (ex_radius in 1:10) { #choose ex values
    
    tryCatch({
      result <- PredictNonlinear(
        dataFrame = dataFrame,
        columns = columns,
        target = target,
        E = E,
        tau = -tau_i,  
        Tp = tau_i,
        exclusionRadius = ex_radius,
        lib = lib_ins,
        pred = lib_oos,
        showPlot = FALSE
      )
      
      # Add exclusionRadius info
      result$exclusionRadius <- ex_radius
      
      # Store result
      results_list[[ex_radius]] <- result
      
    }, error = function(e) {
      # If error happens, store NA and print warning
      message(sprintf("Error at exclusionRadius = %d: %s", ex_radius, e$message))
      results_list[[ex_radius]] <- NA
    })
    
  }
  
  # Remove any NA results
  results_list_clean <- results_list[!sapply(results_list, is.na)]
  
  # Combine into a single dataframe
  if (length(results_list_clean) > 0) {
    results_select_ex <- do.call(rbind, results_list_clean)
  } else {
    results_select_ex <- data.frame()
  }
  
  return(results_select_ex)
}

eval_ex_radius_results <- eval_ex_radius(
  dataFrame = ag_df_CR,
  columns = "phy",
  target = "phy",
  E = 5,
  tau_i = tau_i,
  lib_ins = lib_ins,
  lib_oos = lib_oos
)

simplexE_ex <- print(eval_ex_radius_results[which.max(eval_ex_radius_results$rho), ])

simplexE_ex <- simplexE_ex$rho


#### Smap - Make Predictions 7 DAYS PHY


```{r}

#PREDICT - SMAP

smap_pred <- SMap(dataFrame = ag_df_CR, columns = "phy", target = "chl",
lib = lib_ins, pred = lib_oos, E=simplexE,theta= simplextheta, tau = -tau_i, Tp = tau_i)

 smap_results <- as.data.frame(smap_pred[["predictions"]])


smapstats <- ComputeError(
                          smap_results$Observations,
                       smap_results$Predictions)

#RESULT: 

max_smap_phy_tau7 <- print(smapstats$rho)


chl_smap_pred_tau7 <- smap_pred
chl_max_smap_phy_tau7 <- chl_max_smap_phy_tau7


```
